% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abf2_load.R
\name{abf2_loadlist}
\alias{abf2_loadlist}
\title{abf2_loadlist}
\usage{
abf2_loadlist(filelist, folder = "", attach_ext = TRUE,
  titlelist = NULL)
}
\arguments{
\item{titlelist}{}
}
\description{
abf2_loadlist
}
