% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_base.R
\name{abf_plot_cd}
\alias{abf_plot_cd}
\title{Plot abf objects in channel domain.}
\usage{
abf_plot_cd(abf, intv = NULL, channel = c(2L, 1L), map_func = "mean",
  pack_args = FALSE, ..., colour = TRUE)
}
\arguments{
\item{abf}{an abf object/list of abf objects.}

\item{intv}{a TIME intv to sample abf.}

\item{channel}{channels to map, a two element integer vector, 1st element corresponds to x and 2nd corresponds to y.}

\item{map_func}{a function to map on abf.}

\item{pack_args}{whether to pack arguments for map_func().}

\item{...}{passed to map_func()}

\item{colour}{wheter to plot in coloured mode.}
}
\value{
a ggplot object.
}
\description{
Plot abf objects in channel domain.
}
\details{
This is a low-level plotting function that returns a ggplot object
with plotting data generated by Wrap*() and aesthetics properly mapped.

Since this is a low-level function, there is no data validation or sanity
check performed. Use with caution.
}
