% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_base.R
\name{abf_plot_td}
\alias{abf_plot_td}
\title{Plot abf objects in time domain.}
\usage{
abf_plot_td(abf, intv = NULL, channel = 1L, sample_ratio = 1L,
  sample_func = "mean", sample_colFunc = NULL, ...,
  time_unit = "tick", colour = TRUE)
}
\arguments{
\item{abf}{an abf object/list of abf objects.}

\item{intv}{an INDEX intv to sample abf.}

\item{channel}{channels to map.}

\item{sample_ratio}{sample ratio.}

\item{sample_func}{sample function.}

\item{sample_colFunc}{a sample column function.}

\item{...}{passed to sample_func().}

\item{time_unit}{a time unit passed to TickToTime().}

\item{colour}{whether to plot in coloured mode.}
}
\value{
a ggplot object.
}
\description{
Plot abf objects in time domain.
}
\details{
This is a low-level plotting function that returns a ggplot object
with plotting data generated by MeltAbf() and aesthetics properly mapped.

If only one channel is provided, a "wide" format data.frame is generated and
y is mapped to the channel. If multiple channel is provided, a "long" data.frame
is generated and y is mapped to "value" column.

Since this is a low-level function, there is no data validation or sanity
check performed. Use with caution.
}
