% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{abfapply}
\alias{abfapply}
\title{Apply a function over a list of abf objects.}
\usage{
abfapply(abf_list, FUN, ..., channel = 1L, intv = NULL,
  gen_names = FALSE)
}
\arguments{
\item{abf_list}{a list of abf objects.}

\item{FUN}{the function to be applied to each episode of the abf objects.}

\item{...}{further arguments passed to FUN.}

\item{channel}{channel of the abf objects to apply.}

\item{intv}{OPTIONAL, the interval to apply FUN.}

\item{gen_names}{whether to generate row names.}
}
\value{
a matrix, of which each row represents an episode and column represents an abf object in the list.
}
\description{
Apply a function over a list of abf objects.
}
