% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{AbfPlotChannel}
\alias{AbfPlotChannel}
\alias{AbfPeekChannel}
\title{Plot channel data of abf objects.}
\usage{
AbfPlotChannel(abf, intv = NULL, channel = 1L, cursor = NULL,
  time_unit = "tick", colour = length(channel) > 1L,
  auto_zoom = FALSE, title = NULL, ...)

AbfPeekChannel(abf, intv = NULL, channel = 1L, sample_ratio = 100L,
  cursor = NULL, time_unit = "tick", colour = TRUE,
  auto_zoom = FALSE, title = NULL, ...)
}
\arguments{
\item{abf}{an abf object/a list of abf objects}

\item{intv}{an INDEX intv to sample abf.}

\item{channel}{channels to plot.}

\item{cursor}{add indicative cursors to plot.}

\item{time_unit}{time unit for X axis.}

\item{colour}{wheter to plot in coloured mode.}

\item{auto_zoom}{wheter to zoom into area focused by cursor.}

\item{title}{add title to the plot.}

\item{...}{passed to MeltAbf().}

\item{sample_ratio}{used in AbfPeekChannel, to reduce data points by sample_ratio.}
}
\value{
a ggplot object.
}
\description{
AbfPlotChannel() is a unified wrapper to abf_plot_td(), it is
capable of plotting multiple abf object facetted by id, multiple channels
facetted by channel.

When printing the returned plot with a large number of abf objects, it can be
very slow. It is recommended to use AbfPeekChannel() for previewing and plot
tuning.
}
