% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{AbfPlotXY}
\alias{AbfPlotXY}
\alias{AbfPlotIV}
\title{Plot channel vs channel data of abf objects.}
\usage{
AbfPlotXY(abf, intv = NULL, x_channel = 2L, y_channel = 1L,
  map_func = "mean", colour = TRUE, title = NULL,
  legend_title = "", zero_axes = FALSE, ...)

AbfPlotIV(abf, intv = NULL, map_func = "mean", colour = TRUE,
  title = NULL, legend_title = "", zero_axes = TRUE, ...)
}
\arguments{
\item{abf}{an abf object/a list of abf objects}

\item{intv}{an intv to evaluate channel values.}

\item{x_channel}{channel of x axis.}

\item{y_channel}{channel of y axis.}

\item{map_func}{a function to evaluate channel values.}

\item{colour}{whether to plot in colour mode.}

\item{title}{add title to the plot.}

\item{legend_title}{add legend title to the plot.}

\item{zero_axes}{whether to add zeroed XY axes to the plot.}

\item{...}{passed to map_func()}
}
\value{
a ggplot object.
}
\description{
Plot channel vs channel data of abf objects.
}
