% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_base.R
\name{apply_colFunc_unsafe}
\alias{apply_colFunc_unsafe}
\title{An unsafe apply() that applies a column function}
\usage{
apply_colFunc_unsafe(x, along, colFunc_, keep_dim = FALSE, ...)
}
\arguments{
\item{x}{an array.}

\item{along}{dimension to apply along.}

\item{colFunc_}{a column function to apply.}

\item{keep_dim}{keep dim order}

\item{...}{passed to colFunc_}
}
\value{
an array
}
\description{
Same as apply_unsafe() but applies a column function. The same funtionality
can be achieved by aperm(...) %>% apply(...), however, apply_colFunc_unsafe()
permutates and set dim only once for data to reduce memory footprint.
}
