% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_base.R
\name{apply_unsafe}
\alias{apply_unsafe}
\title{An unsafe apply()}
\usage{
apply_unsafe(x, along, func_, ...)
}
\arguments{
\item{x}{an array.}

\item{along}{dimension to apply along.}

\item{func_}{a function to apply.}

\item{...}{passed to func_}
}
\value{
an array.
}
\description{
Basically apply() without validating anything.
}
