% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_components.R
\name{CursorX}
\alias{CursorX}
\alias{CursorY}
\alias{ApplyCursor}
\title{Apply indicative cursors on plot.}
\usage{
CursorX(cursor, colour = "black")

CursorY(cursor, colour = "black")

ApplyCursor(p, cursor_x = NULL, cursor_y = NULL, colour = "black")
}
\arguments{
\item{cursor}{a numeric vector.}

\item{p}{a ggplot object}

\item{cursor_x}{a numeric vector.}

\item{cursor_y}{a numeric vector.}
}
\value{
a ggplot object.
}
\description{
Apply indicative cursors on plot.
}
