% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveform.R
\name{AtchWaveform}
\alias{AtchWaveform}
\title{Attach a waveform channel to an abf object.}
\usage{
AtchWaveform(abf, dac = GetWaveformEnabledDAC(abf))
}
\arguments{
\item{abf}{an abf object.}

\item{dac}{waveform DAC channel, 1-based.}
}
\value{
an abf object with a new waveform channel attached.
}
\description{
Attach a waveform channel to an abf object.
}
