% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{baseline_als}
\alias{baseline_als}
\title{Baseline Correction with Asymmetric Least Squares Smoothing}
\usage{
baseline_als(y, lambda = 1e+06, p = 0.05, maxitr = 10L,
  converge_warning = FALSE)
}
\arguments{
\item{y}{a numeric vector, signals.}

\item{lambda}{a large numeric, determines smoothness.}

\item{p}{a numeric between (0, 1), determines asymeetry.}

\item{maxitr}{an integer, maximum iteration.}

\item{converge_warning}{whether to give a warning if not converged.}
}
\value{
a numeric vector, baseline of y
}
\description{
Baseline Correction with Asymmetric Least Squares Smoothing
}
\references{
Paul H. C. Eilers, Hans F.M. Boelens, 2005
}
