% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{BaselineEpoch}
\alias{BaselineEpoch}
\title{Calculate baselines of an abf object.}
\usage{
BaselineEpoch(abf, epoch, episodes, channel = 1L, algo = "als", ...)
}
\arguments{
\item{abf}{an abf object.}

\item{epoch}{the epoch name/id to evaluate baselines in.}

\item{episodes}{OPTIONAL, episodes/sweeps to calculate.}

\item{channel}{OPTIONAL, channel id, 1-based.}

\item{algo}{OPTIONAL, algorithm to calculate baselines.}

\item{...}{other arguments to pass to the selected algorithm.}
}
\value{
baselines of selected episodes/sweeps in a list of vectors
}
\description{
Currently only als (Asymmetric Least Squares Smoothing) is available.
}
