% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{BaselineIntv}
\alias{BaselineIntv}
\title{Calculate baselines of an abf object.}
\usage{
BaselineIntv(abf, intv, episodes, channel = 1L, algo = "als", ...)
}
\arguments{
\item{abf}{an abf object.}

\item{intv}{the interval to evaluate baselines in.}

\item{episodes}{episodes/sweeps to calculates.}

\item{channel}{channel id, 1-based.}

\item{algo}{algorithm to calculate baselines.}

\item{...}{other arguments to pass to the selected algorithm.}
}
\value{
baselines of selected episodes/sweeps in a named column matrix
}
\description{
Currently only als (Asymmetric Least Squares Smoothing) is available.
}
