% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{ChannelBaseline}
\alias{ChannelBaseline}
\title{Calculate baseline of a channel}
\usage{
ChannelBaseline(abf, episode = GetAllEpisodes(abf), channel = 1L,
  epoch = NULL, ...)
}
\arguments{
\item{abf}{an abf object.}

\item{episode}{episodes to calculate baseline}

\item{channel}{channel to calculate baseline}

\item{epoch}{OPTIONAL, specific epoch to calculate baseline, if NULL all epochs are calculated.}

\item{...}{arguments passed to baseline_als}
}
\value{
baseline of the channel
}
\description{
Calculate baseline of a channel
}
