% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abf2_episodic_tricks.R
\name{ChannelIntervalMeans}
\alias{ChannelIntervalMeans}
\title{ChannelIntervalMeans calculates interval means of a list of abf objects}
\usage{
ChannelIntervalMeans(abf_list, intv_list, channel = 1, na.rm = TRUE)
}
\arguments{
\item{na.rm}{}
}
\description{
ChannelIntervalMeans calculates interval means of a list of abf objects
}
