% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abf2_episodic_tricks.R
\name{ChannelIntervalSds}
\alias{ChannelIntervalSds}
\title{ChannelIntervalSds calculates interval standard deviations of a list of abf objects}
\usage{
ChannelIntervalSds(abf_list, intv_list, chan_id, na.rm = TRUE)
}
\arguments{
\item{na.rm}{}
}
\description{
ChannelIntervalSds calculates interval standard deviations of a list of abf objects
}
