% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abf2_episodic_tricks.R
\name{ChannelIntervalSems}
\alias{ChannelIntervalSems}
\title{ChannelIntervalSems calculates interval SEM of a list of abf objects}
\usage{
ChannelIntervalSems(abf_list, intv_list, channel = 1, na.rm = TRUE)
}
\arguments{
\item{na.rm}{}
}
\description{
ChannelIntervalSems calculates interval SEM of a list of abf objects
}
