% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ChannelMeanLong}
\alias{ChannelMeanLong}
\title{Calculate mean values of multiple abf objects and return in long format.}
\usage{
ChannelMeanLong(abf_list, intv_list = NULL, channel = NULL,
  group = NULL, na.rm = TRUE)
}
\arguments{
\item{abf_list}{a list of abf objects.}

\item{intv_list}{a list of mean intervals.}

\item{channel}{channels to calculate.}

\item{group}{additional grouping info added to columns.}

\item{na.rm}{whether to remove NAs.}
}
\value{
data.frame
}
\description{
Calculate mean values of multiple abf objects and return in long format.
}
