% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ChannelMeanWide}
\alias{ChannelMeanWide}
\title{Calculate mean values of multiple abf objects and returns in wide format.}
\usage{
ChannelMeanWide(abf_list, intv_list = NULL, channel = 1L,
  na.rm = TRUE)
}
\arguments{
\item{abf_list}{a list of abf objects.}

\item{intv_list}{OPTIONAL, a list of intervals.}

\item{channel}{channel id, 1-based.}

\item{na.rm}{wheter to remove na values..}
}
\value{
A data.frame object/list of data.frame if multiple channels are used.
}
\description{
Calculate mean values of multiple abf objects and returns in wide format.
}
