% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autointv.R
\name{CheckSamplingInterval}
\alias{CheckSamplingInterval}
\title{Print a quick waveform channel summary of a sampling interval.}
\usage{
CheckSamplingInterval(abf, intv, channel, epoch = "B")
}
\arguments{
\item{abf}{an abf object.}

\item{intv}{an interval to summarise.}

\item{channel}{OPTIONAL, the channel to compare.}

\item{epoch}{OPTIONAL, the epoch to compare.}
}
\description{
Print a quick waveform channel summary of a sampling interval.
}
