% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colSds}
\alias{colSds}
\title{Calculate column standard deviations}
\usage{
colSds(df, na.rm = FALSE)
}
\arguments{
\item{df}{a data frame or data frame like 2d data}

\item{na.rm}{remove NAs}
}
\value{
calculated column sd
}
\description{
Calculate column standard deviations
}
