% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colSems}
\alias{colSems}
\alias{rowSems}
\title{Calculate column/row standard error of measurements}
\usage{
colSems(x, na.rm = FALSE, ...)

rowSems(x, na.rm = FALSE, ...)
}
\arguments{
\item{x}{an 2d numeric}

\item{na.rm}{remove NAs}

\item{...}{passed to colSds/rowSds()}
}
\value{
calculated column sem
}
\description{
Calculate column/row standard error of measurements
}
