% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{CropValue}
\alias{CropValue}
\title{Crop undesired values from an abf object}
\usage{
CropValue(abf, channel, max_value, min_value)
}
\arguments{
\item{abf}{an abf object.}

\item{channel}{the channel to crop values.}

\item{max_value}{OPTIONAL, max allowed value}

\item{min_value}{OPTIONAL, min allowed value}
}
\value{
an abf object with values cropped.
}
\description{
All values larger than max_value or smaller than min_value are replaced by NA
during the process, which helps better plotting and ruling out undesired features
such as sudden peak. Use with caution.
}
