% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{DefaultChanLabel}
\alias{DefaultChanLabel}
\title{Get default channel label for every channel of an abf object.}
\usage{
DefaultChanLabel(abf)
}
\arguments{
\item{abf}{an abf object}
}
\value{
a vector of characters.
}
\description{
Get default channel label for every channel of an abf object.
}
