% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{DefaultEpiLabel}
\alias{DefaultEpiLabel}
\title{Get default label for every episode of an abf object}
\usage{
DefaultEpiLabel(abf)
}
\arguments{
\item{abf}{an abf object}
}
\value{
a vector of characters.
}
\description{
Get default label for every episode of an abf object
}
