% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_components.R
\name{DefaultLabeller}
\alias{DefaultLabeller}
\title{Get default facet labeller for an abf object.}
\usage{
DefaultLabeller(abf)
}
\arguments{
\item{abf}{an abf object.}
}
\value{
a labeller.
}
\description{
Get default facet labeller for an abf object.
}
