% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise.R
\name{DenoiseEpoch}
\alias{DenoiseEpoch}
\title{Denoise an abf object by epoch}
\usage{
DenoiseEpoch(abf, epoch, episodes, channel = 1, algo = "sureshrink",
  ...)
}
\arguments{
\item{abf}{an abf object.}

\item{epoch}{epoch name/id.}

\item{episodes}{episodes/sweeps to denoise.}

\item{channel}{channel id, 1-based.}

\item{algo}{algorithm to denoise.}

\item{...}{other arguments to pass to the selected algorithm.}
}
\value{
a named list of denoised episodes/sweeps.
}
\description{
Currently available algorithms are: wavshrink (hard threshold), sureshrink
(soft threshold)
}
