% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abf2_episodic_s3fun.R
\name{dim.ABF2.Episodic}
\alias{dim.ABF2.Episodic}
\title{Dimensions of an abf object}
\usage{
\method{dim}{ABF2.Episodic}(x)
}
\arguments{
\item{x}{an abf object}
}
\description{
Returns the dimension of an abf object.
}
