% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise.R
\name{DnAbf}
\alias{DnAbf}
\alias{DenoiseAbf}
\title{Denoise an abf object by applying wavelet shrink}
\usage{
DnAbf(abf, episode = GetAllEpisodes(abf), channel = 1L,
  by_epoch = TRUE, dac = GetWaveformEnabledDAC(abf),
  thresh.scale = 1, xform = "dwt", ...)

DenoiseAbf(abf, episode, channel = 1L, by_epoch = TRUE, dac,
  thresh.scale = 1, xform = "dwt", ...)
}
\arguments{
\item{abf}{an abf object.}

\item{episode}{episode to denoise.}

\item{channel}{channel to denoise.}

\item{by_epoch}{wheter to denoise by epoch.}

\item{dac}{DAC channel that defines epochs.}

\item{thresh.scale}{controls strength of denoising. See \code{\link[wmtsa:wavShrink]{wavShrink()}} for details.}

\item{xform}{waveform transform type. See \code{\link[wmtsa:wavShrink]{wavShrink()}} for details.}

\item{...}{passed to wavShrink()}
}
\value{
an abf object.
}
\description{
Denoise an abf object by applying wavelet shrink
}
