% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{EpisodeBaseline}
\alias{EpisodeBaseline}
\title{Calculate baseline of an episode}
\usage{
EpisodeBaseline(abf, episode, channel, by_epoch = TRUE,
  dac = GetWaveformEnabledDAC(abf), ...)
}
\arguments{
\item{abf}{an abf object.}

\item{episode}{an episode to calculate baseline}

\item{channel}{channel to calculate baseline}

\item{by_epoch}{calculate baselines by epochs, if set to FALSE, the whole episode
is used to calculate baselines, otherwise, baselines are calculated epoch-by-epoch
then concatenated.}

\item{dac}{dac channel if by_epoch is used.}

\item{...}{arguments passed to baseline_als, see \code{\link[abftools:baseline_als]{baseline_als()}} for details.}
}
\value{
baseline of the channel
}
\description{
Calculate baseline of an episode
}
