% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodic.R
\name{EpisodicIntervalSems}
\alias{EpisodicIntervalSems}
\title{EpisodicIntervalSems calculates interval SEM of a list of abf objects}
\usage{
EpisodicIntervalSems(abf_list, intv_list, channel = 1, na.rm = TRUE)
}
\arguments{
\item{na.rm}{}
}
\description{
EpisodicIntervalSems calculates interval SEM of a list of abf objects
}
