% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuickPlot_aux.R
\name{ExtractDataFrame}
\alias{ExtractDataFrame}
\title{Extract from a data.frame}
\usage{
ExtractDataFrame(df, dict, match_func = startsWith, lower_case = TRUE)
}
\arguments{
\item{df}{data.frame}

\item{dict}{named list}

\item{match_func}{a functio to match}

\item{lower_case}{whether to lower cases of colnames}
}
\value{
a data.frame
}
\description{
Extract columns from a data.frame by matching colnames from a named list. See
ParseDataFrameIV for example usage.
}
\keyword{internal}
