% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_base.R
\name{find_colFunc}
\alias{find_colFunc}
\title{Match function name to a corresponding column functions known to abftools.}
\usage{
find_colFunc(func_name = c("all", "any", "anyNA", "count", "cummax",
  "cummin", "cumprod", "cumsum", "diff", "iqrDiff", "iqr", "logSumExp",
  "madDiff", "mad", "max", "mean", "median", "min", "orderStats", "prod",
  "quantile", "range", "rank", "sdDiff", "sd", "sum", "tabulate",
  "varDiff", "var", "weightedMad", "weightedMean", "weightedMedian",
  "weightedSd", "weightedVar", "sem"))
}
\arguments{
\item{func_name}{a charater vector, function name.}
}
\value{
a function, or NULL if no column function was found.
}
\description{
Match function name to a corresponding column functions known to abftools.
}
\examples{
f <- find_colFunc("mean") #matrixStats::colMeans2
f <- find_colFunc("cos") #NULL, since no column vectorised cos is known to abftools.
}
