% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abf2_episodic_tricks.R
\name{FindAllSamplingInterval}
\alias{FindAllSamplingInterval}
\title{Title}
\usage{
FindAllSamplingInterval(abf_list, current_chan_id = 0,
  voltage_chan_id = 0, min_sampling_size = 0, max_sampling_size = 0,
  allowed_voltage_delta = 0, epoch_name = "auto",
  backward_search = TRUE, strict_comp = TRUE)
}
\arguments{
\item{strict_comp}{}
}
\description{
Title
}
