% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autointv.R
\name{FindChargeInterval}
\alias{FindChargeInterval}
\alias{FindDischargeInterval}
\title{Find a charging interval of given current channel.}
\usage{
FindChargeInterval(abf, epoch = FindMemtestEpoch(abf, dac = dac, type =
  "step"), dac = GetWaveformEnabledDAC(abf),
  episode = FindStepEpisode(abf, epoch = epoch, dac = dac),
  current_channel = GetFirstCurrentChan(abf), peak_window = 3L,
  diff_window = 5L, dy = NULL)

FindDischargeInterval(abf, epoch = FindMemtestEpoch(abf, dac = dac, type
  = "step"), dac = GetWaveformEnabledDAC(abf),
  episode = FindStepEpisode(abf, epoch = epoch, dac = dac),
  current_channel = GetFirstCurrentChan(abf), peak_window = 3L,
  diff_window = 5L, dy = NULL)
}
\arguments{
\item{abf}{an abf object.}

\item{epoch}{epoch to locate the interval.}

\item{dac}{DAC channel of waveform.}

\item{episode}{an episode to locate the interval.}

\item{current_channel}{current channel.}

\item{peak_window}{a window to locate peak.}

\item{diff_window}{a window to calculate slope.}

\item{dy}{1st derivatives of current channel within the epoch.}
}
\value{
an interval.
}
\description{
Find a charging interval of given current channel.
}
\details{
This function finds an interval of an episode that fits charging/discharging
behaviour of an RC circuit.
Since abf object records descrete data points, FindChargeInterval() may have
numeric stability issues in some extreme cases. Increase peak_window and
diff_window usually solves them. The default values of peak_window, diff_window
are tested and should work most of the time.
You can also provide a pre-calculated 1st derivative using argument dy. Please
note that dy should only contains points within the epoch instead of a whole
sweep.
}
