% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epoch.R
\name{FindMemtestEpoch}
\alias{FindMemtestEpoch}
\title{Look for first epoch that can perform a membrane test.}
\usage{
FindMemtestEpoch(abf, dac = GetWaveformEnabledDAC(abf),
  type = c("step", "ramp"))
}
\arguments{
\item{abf}{an abf object.}

\item{dac}{dac channel.}

\item{type}{type of membrane test. When type is "step", the function searches
for pattern of holding level -> step -> holding level. When type is "ramp",
the function searches for pattern of holding level -> ramp up/down -> ramp down/up
-> holding level.}
}
\value{
an epoch number
}
\description{
Look for first epoch that can perform a membrane test.
}
