% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autointv.R
\name{FindSamplingInterval}
\alias{FindSamplingInterval}
\alias{FindAllSamplingInterval}
\title{FindSamplingInterval finds a stable interval for sampling current and voltage
channel of an abf object.}
\usage{
FindSamplingInterval(abf, epoch = NULL,
  dac = GetWaveformEnabledDAC(abf),
  current_channel = GetFirstCurrentChan(abf),
  voltage_channel = GetFirstVoltageChan(abf),
  target_interval_size = NULL, allowed_voltage_delta = NULL,
  noisy_opt = FALSE, lp_freq = 75, lp_order = 1L)

FindAllSamplingInterval(abf_list, ...)
}
\arguments{
\item{abf}{an abf object.}

\item{epoch}{the epoch to search, defaults to first multi-step epoch.}

\item{dac}{the dac channel to evaluate waveform.}

\item{current_channel}{channel id for current data, channel id is 1-based.}

\item{voltage_channel}{channel id for voltage data, channel id is 1-based.}

\item{target_interval_size}{OPTIONAL, target size in **points** of the sampling interval.
Default is 2x lowpass window.}

\item{allowed_voltage_delta}{OPTIONAL, allowed max deviation of voltage.}

\item{noisy_opt}{enable optimisation for noisy data by applying a Butterworth
lowpass filter to current channel.}

\item{lp_freq}{frequency of low-pass filter.}

\item{lp_order}{order of low-pass filter.}
}
\value{
a named vector of 3 numeric: interval start position, end position, length
}
\description{
FindSamplingInterval finds a stable interval for sampling current and voltage
channel of an abf object.
}
\details{
This function finds a proper sampling interval to perform further
calculations in a multi-step TEVC abf object. The waveform epochs should be
aligned, i.e. epoch duration incr is 0. The function finds a proper sampling
interval by:

1. Calling CmpWaveform() to find stable VOLTAGE channel, thus minimizing SEM
along V axis.

2. Performing a binary search on CURRENT channel to find the most stable interval.

3. Applying a time-dependent penalty function to make the algorithm prefer
intervals that are closer to the end of the epoch.
}
