% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autointv.R
\name{FindStepEpisode}
\alias{FindStepEpisode}
\title{Find a step episode.}
\usage{
FindStepEpisode(abf, epoch = FindMemtestEpoch(abf, dac = dac, type =
  "step"), dac = GetWaveformEnabledDAC(abf))
}
\arguments{
\item{abf}{an abf object.}

\item{epoch}{epoch id.}

\item{dac}{DAC channel of waveform.}
}
\value{
an episode id.
}
\description{
This function compares the DAC level of given epoch to previous epoch, and find
the episode with smallest step size.
}
