% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacitance.R
\name{fit_charge_cap}
\alias{fit_charge_cap}
\title{Fit a charging curve and calculate Rs, Rm, Cm}
\usage{
fit_charge_cap(i, intv_charge, intv_hold, delta_v, time_unit = "tick",
  sampling_rate)
}
\arguments{
\item{i}{current episode}

\item{intv_charge}{charging interval}

\item{intv_hold}{hold interval}

\item{delta_v}{delta voltage}

\item{time_unit}{time unit to use.}

\item{sampling_rate}{sampling ratio passed to TickToTime()}
}
\value{
a list of fitted results.
}
\description{
Fit a charging curve and calculate Rs, Rm, Cm
}
