% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacitance.R
\name{fit_charge_i}
\alias{fit_charge_i}
\title{Fit a current curve to a logistic aysmptotic model.}
\usage{
fit_charge_i(i, t)
}
\arguments{
\item{i}{current}

\item{t}{time}
}
\value{
a list of fitted coefs, and fitted model itself.
}
\description{
Model:
i = I0 + (Is - I0) * e^( -exp(lrc) * t )
}
