% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_components.R
\name{ForceZoomY}
\alias{ForceZoomY}
\alias{ApplyForceZoomY}
\title{Force zooming into Y axis.}
\usage{
ForceZoomY(xrange, yrange = NULL, xdata, ydata, space = 0.0125,
  clipping = FALSE)

ApplyForceZoomY(p, xrange, yrange = NULL, space = 0.0125,
  clipping = FALSE)
}
\arguments{
\item{xrange}{Zoomed range of x axis.}

\item{yrange}{Zoomed range of y axis. If NULL, ydata will be evaluated.}

\item{xdata}{Data of x.}

\item{ydata}{Data of y, indices correspond to xdata.}

\item{space}{Spacing factor.}

\item{clipping}{Whether to "throw away" out of limit data.}

\item{p}{A ggplot object.}
}
\value{
ylim
}
\description{
Force zooming into Y axis.
}
\details{
When clippling is TRUE, ylim() is applied directly to p, causing
out-of-limit data to be removed.
}
