% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aux.R
\name{FreqToTick}
\alias{FreqToTick}
\title{Calculate time span of a frequency}
\usage{
FreqToTick(abf, freq, sampling_rate)
}
\arguments{
\item{abf}{an abf object.}

\item{freq}{a frequency in Hz.}

\item{sampling_rate}{sampling rate (in µs).}
}
\value{
an integer of **point**/**tick** count.
}
\description{
Calculate time span of a frequency
}
