% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_attr.R
\name{GetAllChannels}
\alias{GetAllChannels}
\title{Return all channels of an abf object.}
\usage{
GetAllChannels(abf)
}
\arguments{
\item{abf}{an abf object.}
}
\value{
a vector of channel ids.
}
\description{
Return all channels of an abf object.
}
