% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_attr.R
\name{GetAllEpisodes}
\alias{GetAllEpisodes}
\title{Return all episodes of an abf object.}
\usage{
GetAllEpisodes(abf)
}
\arguments{
\item{abf}{an abf object.}
}
\value{
a vector of episode numbers.
}
\description{
Return all episodes of an abf object.
}
