% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode.R
\name{GetAvailEpisodes}
\alias{GetAvailEpisodes}
\title{Get available episodes}
\usage{
GetAvailEpisodes(abf)
}
\arguments{
\item{abf}{an abf or a list of abf object.}
}
\value{
episodes that are not marked removed.
}
\description{
Get available episodes
}
