% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{GetBaseline}
\alias{GetBaseline}
\title{Calculate baselines of an abf object}
\usage{
GetBaseline(abf, epoch, intv, episodes = 0, channel = 1,
  algo = "als", ...)
}
\arguments{
\item{abf}{an abf object}

\item{epoch}{an epoch name/id to of which the baselines will be calculated}

\item{intv}{an interval in which the baselines will be calculated}

\item{episodes}{episodes to calculate}

\item{channel}{channel id, 1-based}

\item{algo}{algorithm to calculate baselines}

\item{...}{other arguments to pass to the algorithm}
}
\value{
a list of vectors
}
\description{
Currently only als (Asymmetric Least Squares Smoothing) is available.
}
