% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{GetChanLabel}
\alias{GetChanLabel}
\title{Get channel label for every channel of an abf object.}
\usage{
GetChanLabel(abf, style = "\%s (\%s)")
}
\arguments{
\item{abf}{an abf object}

\item{style}{a format string}
}
\value{
a vector of characters.
}
\description{
Get channel label for every channel of an abf object.
}
