% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R
\name{GetChannelDesc}
\alias{GetChannelDesc}
\title{Get descriptions of channels.}
\usage{
GetChannelDesc(abf, channel = NULL)
}
\arguments{
\item{abf}{an abf object.}
}
\value{
a character vector of channel descriptions.
}
\description{
Get descriptions of channels.
}
