% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R
\name{GetChannelUnit}
\alias{GetChannelUnit}
\title{Get units of channels.}
\usage{
GetChannelUnit(abf, channel = NULL)
}
\arguments{
\item{abf}{an abf object.}

\item{channel}{channel id.}
}
\value{
a character vector of channel units.
}
\description{
Get units of channels.
}
