% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_attr.R
\name{GetCurrentChan}
\alias{GetCurrentChan}
\title{Get current channel id.}
\usage{
GetCurrentChan(abf)
}
\arguments{
\item{abf}{an abf object or a list of abf objects.}
}
\value{
a vector of current channel id.
}
\description{
Get current channel id.
}
