% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aux.R
\name{GetEpiLabel}
\alias{GetEpiLabel}
\alias{GetChanLabel}
\alias{DefaultEpiLabel}
\alias{DefaultChanLabel}
\title{Compose labels for every episode/channel of an abf object}
\usage{
GetEpiLabel(abf, style = "epi\%d")

GetChanLabel(abf, style = "\%s (\%s)")

DefaultEpiLabel(abf)

DefaultChanLabel(abf)
}
\arguments{
\item{abf}{an abf object}

\item{style}{a format string}
}
\value{
a vector of characters.
}
\description{
Compose labels for every episode/channel of an abf object
}
