% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R
\name{GetEpisodesPerChannel}
\alias{GetEpisodesPerChannel}
\alias{nEpi}
\title{Get number of episodes/sweeps per channel.}
\usage{
GetEpisodesPerChannel(abf)

nEpi(abf)
}
\arguments{
\item{abf}{an abf object.}
}
\value{
number of episodes/sweeps per channel.
}
\description{
Get number of episodes/sweeps per channel.
}
