% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aux.R
\name{GetEpiTag}
\alias{GetEpiTag}
\alias{GetChanTag}
\title{Compose default episode/channel number tag (used as column names).}
\usage{
GetEpiTag(episode)

GetChanTag(channel)
}
\arguments{
\item{episode}{episode numbers}

\item{channel}{channel numbers}
}
\value{
a vector of characters
}
\description{
Compose default episode/channel number tag (used as column names).
}
\examples{
epi_tag <- GetEpiTag(c(1,3,4,5))
ch_tag <- GetChanTag(2:3)
}
