% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epoch.R
\name{GetEpochIntervals}
\alias{GetEpochIntervals}
\title{Get intervals of all epochs.}
\usage{
GetEpochIntervals(abf, dac = GetWaveformEnabledDAC(abf))
}
\arguments{
\item{abf}{an abf object.}

\item{dac}{OPTIONAL, DAC channel, 1-based.}
}
\value{
a 3-d array, see details.
}
\description{
The returned array contains all interval info, which can be accessed by:
epoch[ , epoch_id, episode], an interval is defined as c(intv_start, intv_end, intv_length)
}
