% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_attr.R
\name{GetFirstCurrentChan}
\alias{GetFirstCurrentChan}
\title{Get first current channel id.}
\usage{
GetFirstCurrentChan(abf)
}
\arguments{
\item{abf}{an abf object or a list of abf objects.}
}
\value{
an integer id of first current channel.
}
\description{
Get first current channel id.
}
