% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_attr.R
\name{GetFirstVoltageChan}
\alias{GetFirstVoltageChan}
\alias{GetFirstCurrentChan}
\title{Get first voltage/current channel id.}
\usage{
GetFirstVoltageChan(abf)

GetFirstCurrentChan(abf)
}
\arguments{
\item{abf}{an abf object or a list of abf objects.}
}
\value{
an integer id of first voltage/current channel.
}
\description{
Get first voltage/current channel id.
}
