% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R
\name{GetNumOfChannel}
\alias{GetNumOfChannel}
\title{Get number of channels.}
\usage{
GetNumOfChannel(abf)
}
\arguments{
\item{abf}{an abf object.}
}
\value{
number of channels.
}
\description{
Get number of channels.
}
