% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R
\name{GetNumOfDAC}
\alias{GetNumOfDAC}
\alias{nDAC}
\title{Get number of DAC.}
\usage{
GetNumOfDAC(abf)

nDAC(abf)
}
\arguments{
\item{abf}{}
}
\value{
a vector of integer, number of DAC
}
\description{
Get number of DAC.
}
