% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R
\name{GetNumOfEpoch}
\alias{GetNumOfEpoch}
\alias{nEpoch}
\title{Get number of epochs.}
\usage{
GetNumOfEpoch(abf, dac = 1L)

nEpoch(abf, dac = 1L)
}
\arguments{
\item{abf}{an abf object}

\item{dac}{id of DAC channel}
}
\value{
number of epochs.
}
\description{
Get number of epochs.
}
