% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R
\name{GetPointsPerEvent}
\alias{GetPointsPerEvent}
\alias{nPtsEvent}
\title{Get number of recorded points for corresponding event.}
\usage{
GetPointsPerEvent(abf, event = 1L)

nPtsEvent(abf, event = 1L)
}
\arguments{
\item{abf}{an abf object.}

\item{event}{id of event.}
}
\value{
number of recorded points.
}
\description{
Get number of recorded points for corresponding event.
}
