% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abf2_protocol.R
\name{GetProtocol}
\alias{GetProtocol}
\title{Return the protocols of an abf object.}
\usage{
GetProtocol(abf)
}
\arguments{
\item{abf}{an abf object.}
}
\value{
a list of data frames of protocol sections.
}
\description{
Return the protocols of an abf object.
}
