% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epoch.R
\name{GetStepEpoch}
\alias{GetStepEpoch}
\alias{GetMultiStepEpoch}
\title{Return all step epoch id.}
\usage{
GetStepEpoch(abf, dac = GetWaveformEnabledDAC(abf))

GetMultiStepEpoch(abf, dac = GetWaveformEnabledDAC(abf))
}
\arguments{
\item{abf}{an abf object.}

\item{dac}{DAC channel.}
}
\value{
an integer vector.
}
\description{
Return all step epoch id.
}
