% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R
\name{GetSynchArray}
\alias{GetSynchArray}
\title{Return sync array of an abf object.}
\usage{
GetSynchArray(abf)
}
\arguments{
\item{abf}{an abf object.}
}
\value{
a sync array
}
\description{
Return sync array of an abf object.
}
