% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R
\name{GetTitle}
\alias{GetTitle}
\title{Get title of an abf object.}
\usage{
GetTitle(abf)
}
\arguments{
\item{abf}{an abf object.}
}
\value{
the title of the abf object.
}
\description{
Get title of an abf object.
}
