% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveform.R
\name{GetWaveform}
\alias{GetWaveform}
\title{Simulate waveform.}
\usage{
GetWaveform(abf, episode = GetAllEpisodes(abf),
  dac = GetWaveformEnabledDAC(abf))
}
\arguments{
\item{abf}{an abf object.}

\item{episode}{the episodes to simulate.}

\item{dac}{waveform DAC channel, 1-based.}
}
\value{
channel data of the simulated waveform.
}
\description{
Simulate waveform.
}
