% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epoch.R
\name{GetWaveformEnabledDAC}
\alias{GetWaveformEnabledDAC}
\title{Get DAC id of which waveform is enabled.}
\usage{
GetWaveformEnabledDAC(abf)
}
\arguments{
\item{abf}{an abf object.}
}
\value{
DAC id, 1-based.
}
\description{
Get DAC id of which waveform is enabled.
}
