% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Intv}
\alias{Intv}
\title{Generate an interval.}
\usage{
Intv(startPos, endPos, len)
}
\arguments{
\item{startPos}{start position.}

\item{endPos}{end position.}

\item{len}{length of the interval.}
}
\value{
an interval.
}
\description{
Only two or the arguments are needed. If all are given, len will be ignored.
}
