% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ISummaryWide}
\alias{ISummaryWide}
\alias{VSummaryWide}
\title{Calculate mean current/voltage of multiple abf objects and return in wide format.}
\usage{
ISummaryWide(abf_list, intv_list = NULL,
  current_channel = GetFirstCurrentChan(abf_list), na.rm = TRUE)

VSummaryWide(abf_list, intv_list = NULL,
  voltage_channel = GetFirstVoltageChan(abf_list), na.rm = TRUE)
}
\arguments{
\item{abf_list}{a list of abf objects.}

\item{intv_list}{OPTIONAL, a list of intervals.}

\item{current_channel}{current channel id.}

\item{na.rm}{wheter to remove na values.}

\item{voltage_channel}{voltage channel id.}
}
\value{
A data.frame object.
}
\description{
Calculate mean current/voltage of multiple abf objects and return in wide format.
}
