% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{IVSummary}
\alias{IVSummary}
\title{IVSummary calculates average current of a list of abf objects, within given intervals}
\usage{
IVSummary(abf_list, intv_list, current_channel = 0,
  voltage_channel = 0)
}
\arguments{
\item{voltage_channel}{}
}
\description{
IVSummary calculates average current of a list of abf objects, within given intervals
}
