% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise.R
\name{LowpassAbf}
\alias{LowpassAbf}
\title{Apply Butterworth low-pass filter to an abf object.}
\usage{
LowpassAbf(abf, channel, freq = 75, order = 1L)
}
\arguments{
\item{abf}{an abf object.}

\item{channel}{the channel to apply filter.}

\item{freq}{low-pass frequency.}

\item{order}{filter order.}
}
\value{
an abf object.
}
\description{
Apply Butterworth low-pass filter to an abf object.
}
