% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_base.R
\name{mapnd}
\alias{mapnd}
\alias{mapnd_col}
\title{Mapping function to an nd-array along a specific axis.}
\usage{
mapnd(x, func, along = 1L, pack_args = FALSE, ...)

mapnd_col(x, colFunc, along = 1L, ...)
}
\arguments{
\item{x}{an nd-array.}

\item{func}{a function to map.}

\item{along}{the axis to map the function to (dims to collapse).}

\item{pack_args}{whether to pack arguments for func.}

\item{...}{other arguments passed to func.}

\item{colFunc}{a column function.}

\item{func_name}{a optimised function name.}
}
\value{
an array (dimension depending on returned dimension of func)
}
\description{
mapnd() applies a function along a specific axis/dimension of x. However,
if a corresponding column function is available, using mapnd_col() can be
more efficient e.g. use mapnd_col(x, colMeans) instead of mapnd(x, mean).
For convenience, you can also call mapnd() by supplying a function name, e.g.
mapnd(x, "mean"). In this case, mapnd will first lookup any corresponding
column vectorised function by calling find_colFunc() and if a colFunc is found,
mapnd_col will be used instead for better performance.
}
