% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{mean.abf}
\alias{mean.abf}
\title{Calculate mean values of an abf object.}
\usage{
\method{mean}{abf}(abf, intv = NULL, ret.df = FALSE,
  use_chan_name = FALSE, na.rm = TRUE, ...)
}
\arguments{
\item{abf}{an abf object.}

\item{intv}{OPTIONAL, an interval to mean over.}

\item{ret.df}{whether to return a data.frame object, if set to FALSE a matrix is returned instead.}

\item{use_chan_name}{whether to use channel names as colnames, set to FALSE to use descriptive names instead.}

\item{na.rm}{whether to remove NA values.}

\item{...}{passed to arithmetic mean.}
}
\value{
A matrix/data.frame.
}
\description{
Returns a matrix/data.frame containing episodic mean values of each channel.
}
