% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{melt.abf}
\alias{melt.abf}
\title{Melt channel data of an abf object into a data frame by sampling ratio.}
\usage{
melt.abf(abf, channel = 1L, sampling_ratio = 1L,
  sampling_func = NULL, time_unit = "tick")
}
\arguments{
\item{abf}{an abf object}

\item{channel}{the channel to melt, default to 1}

\item{sampling_ratio}{sampling ratio, default to 1 so that no sampling at all}

\item{sampling_func}{a function, default to null so no processing sampled points}

\item{time_unit}{time unit of the sampled data, can be tick, us, ms or s}
}
\description{
For easier melting abf channel data, sampling time (by given time unit) is
selected as id.vars, variable.name is set to Episode and value.name is
determined by your selected channel, which is extracted from ChannelDesc attr
of the abf object.
}
\details{
You can also provide sampling_ratio and sampling_func to reduce data points
of the molten data frame. A sampling function should take a numeric vector as
input and return a single value or be left NULL so the data is sampled at the
exact position every sampling_ratio points.
}
\examples{
Melt first channel with all default behaviors:
df <- melt(abf)
Melt second channel, sample every 200 points and take their mean, using time unit of ms:
df <- melt(abf, channel = 2, sampling_ratio = 200, sampling_func = mean, time_unit = "ms")
}
