% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuickPlot_aux.R
\name{MeltAbfChannel}
\alias{MeltAbfChannel}
\title{Melt channel data of abf objects.}
\usage{
MeltAbfChannel(abf, channel = NULL, intv = NULL, map_func = mean,
  abf_id_func = GetTitle, epi_id_func = DefaultEpiLabel,
  chan_id_func = DefaultChanLabel, na.intv.rm = TRUE, ...)
}
\arguments{
\item{abf}{an abf/list of abf objects.}

\item{channel}{channel/channels to melt.}

\item{intv}{OPTIONAL, an interval to process.}

\item{map_func}{a mapping function to process data.}

\item{abf_id_func}{OPTIONAL, a function accepts an abf object and returns an identifier of the objects.}

\item{epi_id_func}{OPTIONAL, a function accepts an abf object and returns a vector of identifiers of all episodes.}

\item{chan_id_func}{OPTIONAL, a function accepts an abf object and returns a vectors of identifiers of all channels.}

\item{na.intv.rm}{whether to remove NA intervals passed.}

\item{...}{further arguments passed to map_func.}
}
\value{
a melted data frame
}
\description{
Melt channel data of abf objects.
}
