% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodic.R
\name{MultiIntervalMeans}
\alias{MultiIntervalMeans}
\title{MultiIntervalMeans calculates interval means of a list of abf objects}
\usage{
MultiIntervalMeans(abf_list, intv_list, channel = 1L, na.rm = TRUE)
}
\arguments{
\item{abf_list}{a list of abf objects.}

\item{intv_list}{a list of intervals.}

\item{channel}{channel to calculated, channel id is 1-based.}

\item{na.rm}{remove NAs.}
}
\value{
a vector of calculated values.
}
\description{
MultiIntervalMeans calculates interval means of a list of abf objects
}
