% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodic.R
\name{MultiIntervalSems}
\alias{MultiIntervalSems}
\title{MultiIntervalSems calculates interval SEM of a list of abf objects}
\usage{
MultiIntervalSems(abf_list, intv_list, channel = 1, na.rm = TRUE)
}
\arguments{
\item{abf_list}{a list of abf objects.}

\item{intv_list}{a list of intervals.}

\item{channel}{channel to calculated, channel id is 1-based.}

\item{na.rm}{remove NAs.}
}
\value{
a vector of calculated values.
}
\description{
MultiIntervalSems calculates interval SEM of a list of abf objects
}
