% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{MultiMean_Current}
\alias{MultiMean_Current}
\title{Calculate mean currents of multiple abf objects}
\usage{
MultiMean_Current(abf_list, intv_list, na.rm = TRUE)
}
\arguments{
\item{abf_list}{a list of abf objects.}

\item{intv_list}{a list of intervals.}

\item{na.rm}{wheter to remove na values..}
}
\value{
A data.frame object.
}
\description{
Calculate mean currents of multiple abf objects
}
