% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiPlot.R
\name{MultiPeekChannel}
\alias{MultiPeekChannel}
\title{Title}
\usage{
MultiPeekChannel(abf_list, channel = 1, colour = FALSE,
  time_unit = "tick", num_label = FALSE, title_label = TRUE,
  unify_y = TRUE, ratio = 100, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
