% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiPlot.R
\name{MultiPeekChannel_Intv}
\alias{MultiPeekChannel_Intv}
\title{Fast plot a channel of multiple abf objects, with intervals.}
\usage{
MultiPeekChannel_Intv(abf_list, intv_list, channel = 1, colour = FALSE,
  time_unit = "tick", num_label = FALSE, title_label = TRUE,
  unify_y = TRUE, ratio = 100, ...)
}
\arguments{
\item{abf_list}{a list of abf objects.}

\item{intv_list}{a list of intervals.}

\item{channel}{channel id, 1-based.}

\item{colour}{whether to plot in coloured mode.}

\item{time_unit}{time unit to use in plot.}

\item{num_label}{whether to show a numbered label of each abf object.}

\item{title_label}{wheter to show title of each abf object.}

\item{unify_y}{wheter to unifier y scale of each abf object.}

\item{ratio}{sampling ratio for faster plotting.}

\item{...}{other arguments, see melt.abf for more details.}
}
\value{
a ggplot object.
}
\description{
Fast plot a channel of multiple abf objects, with intervals.
}
