% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiPlot.R
\name{MultiPlotChannel_Intv}
\alias{MultiPlotChannel_Intv}
\title{Title}
\usage{
MultiPlotChannel_Intv(abf_list, intv_list, channel = 1, colour = FALSE,
  time_unit = "tick", num_label = FALSE, title_label = TRUE,
  unify_y = TRUE, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
