% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiPlot.R
\name{MultiPlotChannel_Pub}
\alias{MultiPlotChannel_Pub}
\title{Plot a channel of multiple abf objects, in a common publication standard.}
\usage{
MultiPlotChannel_Pub(abf_list, channel = 1L, time_unit = "s",
  manual_title = NULL, auto_title = FALSE, x_ticks_pos = c(0.2, 0.8),
  ...)
}
\arguments{
\item{abf_list}{a list of abf objects.}

\item{channel}{channel id, 1-based.}

\item{time_unit}{time unit to use in plot.}

\item{manual_title}{titles for subplots.}

\item{auto_title}{automatically add titles for subplots.}

\item{x_ticks_pos}{a vector of positions (valued between 0 and 1) to show x ticks.}

\item{...}{other arguments passed, see PlotChannel for more details.}
}
\value{
a ggplot object.
}
\description{
Plot a channel of multiple abf objects, in a common publication standard.
}
