% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R
\name{nEpi}
\alias{nEpi}
\title{Get number of episodes/sweeps per channel.}
\usage{
nEpi(abf)
}
\arguments{
\item{abf}{an abf object.}
}
\value{
number of episodes/sweeps per channel.
}
\description{
The returned number may be different to the result of dim(abf[[chan_id]]) if
you have removed episodes from the abf object. GetEpisodesPerChannel / nEpi
always return the original number of episodes per channel setting in the
abf2 protocol.
}
