% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{PackArgs}
\alias{PackArgs}
\title{Packing arguments in function call}
\usage{
PackArgs(f, ...)
}
\arguments{
\item{f}{a function}

\item{...}{other arguments passed to f}
}
\value{
a function that accepts a vector argument
}
\description{
This is a helper function that can be handy when mapping function calls to
multiple data. It wraps a function f, which accepts multiple arguments, into
a new function that accept a vector as its argument.
}
\examples{
ivpair <- c(10, 50)
resistance <- function(v, i) v / i
f <- PackArgs(resistance)
r1 <- resistance(ivpair[1], ivpair[2])
r2 <- f(ivpair)

}
