% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak.R
\name{PeakDetectEpoch}
\alias{PeakDetectEpoch}
\title{Peak detection by epoch.}
\usage{
PeakDetectEpoch(abf, epoch, episodes, channel = 1, algo = "cwt", ...)
}
\arguments{
\item{abf}{an abf object.}

\item{epoch}{epoch id/name.}

\item{episodes}{episodes/sweeps to detect peaks.}

\item{channel}{channel id, 1-based.}

\item{algo}{algorithm for peak detection.}

\item{...}{other argumetns to pass to the selected algorithm.}
}
\value{
a named list of detected peaks' x positions.
}
\description{
Currently available algorithms: cwt, zscore
}
