% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak.R
\name{PeakDetectIntv}
\alias{PeakDetectIntv}
\title{Peak detection by epoch.}
\usage{
PeakDetectIntv(abf, intv, episodes = 0, channel = 1, algo = "cwt",
  ...)
}
\arguments{
\item{abf}{an abf object.}

\item{intv}{an interval to detect peaks.}

\item{episodes}{episodes/sweeps to detect peaks.}

\item{channel}{channel id, 1-based.}

\item{algo}{algorithm for peak detection.}

\item{...}{other argumetns to pass to the selected algorithm.}
}
\value{
a named list of detected peaks' x positions.
}
\description{
Currently available algorithms: cwt, zscore
}
