% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotChannel.R
\name{PeekAllChannel_Cursor_H}
\alias{PeekAllChannel_Cursor_H}
\title{Fast all channels with cursors horizontally.}
\usage{
PeekAllChannel_Cursor_H(abf, cursor, colour = FALSE, ratio = 50, ...)
}
\arguments{
\item{abf}{an abf object.}

\item{cursor}{cursors to plot.}

\item{colour}{wheter to plot in coloured mode.}

\item{ratio}{sampling ratio.}

\item{...}{other arguments, see melt.abf for more details.}
}
\value{
a ggplot object.
}
\description{
Fast all channels with cursors horizontally.
}
