% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotChannel.R
\name{PeekAllChannel_H}
\alias{PeekAllChannel_H}
\title{Fast plot all channels horizontally}
\usage{
PeekAllChannel_H(abf, colour = FALSE, ratio = 50, ...)
}
\arguments{
\item{abf}{an abf object.}

\item{colour}{wheter to plot in coloured mode.}

\item{ratio}{sampling ratio}

\item{...}{other arguments, see melt.abf for more details.}
}
\value{
a ggplot object.
}
\description{
Fast plot all channels horizontally
}
