% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotChannel.R
\name{PeekAllChannel_Intv_V}
\alias{PeekAllChannel_Intv_V}
\title{Fast plot all channels with interval vertically.}
\usage{
PeekAllChannel_Intv_V(abf, intv, colour = FALSE, ratio = 50, ...)
}
\arguments{
\item{abf}{an abf object.}

\item{intv}{an interval.}

\item{colour}{wheter to plot in coloured mode.}

\item{ratio}{sampling ratio.}

\item{...}{other arguments, see melt.abf for more details.}
}
\value{
a ggplot object.
}
\description{
Fast plot all channels with interval vertically.
}
