% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotChannel.R
\name{PlotChannel}
\alias{PlotChannel}
\title{Plot a channel of an abf object.}
\usage{
PlotChannel(abf, intv = NULL, cursor = NULL, channel = 1L,
  colour = FALSE, time_unit = "tick", auto_zoom = FALSE,
  title = NULL, ...)
}
\arguments{
\item{abf}{an abf object.}

\item{intv}{OPTIONAL, an interval (a vector of c(start, end, len)) to plot on top.}

\item{cursor}{OPTIONAL, cursors (a vector of positions) to plot on top.}

\item{channel}{channel to plot, 1-based.}

\item{colour}{whether to plot in coloured mode.}

\item{time_unit}{time unit for x axis.}

\item{auto_zoom}{whether to zoom in the plot automatically.}

\item{title}{OPTIONAL, title of the plot.}

\item{...}{other arguments passed to melt, see melt.abf for more details.}
}
\value{
a ggplot object.
}
\description{
Plot a channel of an abf object.
}
