% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotChannelAbf}
\alias{PlotChannelAbf}
\alias{PeekChannelAbf}
\title{Plot channel data of abf objects.}
\usage{
PlotChannelAbf(abf, intv = NULL, channel = GetAllChannels(abf),
  episode = GetAvailEpisodes(abf), concat_epi = FALSE, cursor = NULL,
  time_unit = "tick", colour = max(nEpi(abf)) > 1L,
  auto_zoom = FALSE, title = NULL, ...)

PeekChannelAbf(abf, intv = NULL, channel = 1L, sample_ratio = 100L,
  cursor = NULL, time_unit = "tick", colour = TRUE,
  auto_zoom = FALSE, title = NULL, ...)
}
\arguments{
\item{abf}{an abf object/a list of abf objects}

\item{intv}{an INDEX intv to sample abf.}

\item{channel}{channels to plot.}

\item{episode}{episodes to group.}

\item{concat_epi}{wheter to concatenate all episodes.}

\item{cursor}{add indicative cursors to plot.}

\item{time_unit}{time unit for X axis.}

\item{colour}{wheter to plot in coloured mode.}

\item{auto_zoom}{wheter to zoom into area focused by cursor.}

\item{title}{add title to the plot.}

\item{...}{passed to MeltAbf().}
}
\value{
a ggplot object.
}
\description{
AbfPlotChannel() is a unified wrapper to abf_plot_td(), it is
capable of plotting multiple abf object facetted by id, multiple channels
facetted by channel.

When printing the returned plot with a large number of abf objects, it can be
very slow. It is recommended to use AbfPeekChannel() for previewing and plot
tuning.
}
