% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotIVSummary}
\alias{PlotIVSummary}
\title{Plot IV Summary generated by IVSummary().}
\usage{
PlotIVSummary(ivs, mode = c("i", "g", "spi", "spg"), colour = TRUE,
  title = NULL, legend_title = "", error_bar = FALSE,
  zero_axes = TRUE, smooth = FALSE, ...)
}
\arguments{
\item{ivs}{a data.frame generated by IVSummary or a list of such data.frames.}

\item{mode}{plotting mode, see details for more information.}

\item{colour}{whether to plot in colour mode.}

\item{title}{title of the plot.}

\item{legend_title}{title of the legend (group).}

\item{error_bar}{whether to plot error bars.}

\item{zero_axes}{wheter to apply ZeroAxes() to plot.}

\item{smooth}{wheter to plot b-Spline smoothed lines, see details for more information.}

\item{...}{not used.}
}
\value{
a ggplot object.
}
\description{
Plot IV Summary generated by IVSummary().
}
\details{
plotting mode can be i, g, spi, spg, which stand for current, conductance,
specific current and specific conducatance and the plotting function
looks for Current/Conductance/SpCurrent/SpConductance column respectively.

PlotIVSummary() uses basic splines to achieve smoothing. However, b-spline
smoothing can be misleading especially when I-V follows logistic and hyperbolic
relationship. In these cases, you should always use a suitable fitting formula
to smooth.
}
