% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3func.R
\name{print.abf}
\alias{print.abf}
\title{Print an abf object}
\usage{
\method{print}{abf}(x, ...)
}
\arguments{
\item{x}{an abf object.}

\item{...}{not used.}
}
\value{
the abf object itself, invisibly.
}
\description{
Print an abf object
}
