% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abf2_episodic_s3fun.R
\name{print.ABF2.Episodic}
\alias{print.ABF2.Episodic}
\title{print summary of an abf object}
\usage{
\method{print}{ABF2.Episodic}(x, ...)
}
\arguments{
\item{x}{an abf object}

\item{...}{further arguments for future use}
}
\description{
print summary of an abf object
}
