% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuickPlot.R
\name{QuickPlot}
\alias{QuickPlot}
\alias{QuickPlot.default}
\alias{QuickPlot.matrix}
\alias{QuickPlot.abf}
\alias{QuickPlot.data.frame}
\alias{QuickPlot.list}
\title{Quick plot various type of data.}
\usage{
QuickPlot(data, ...)

\method{QuickPlot}{default}(data, ...)

\method{QuickPlot}{matrix}(data, ...)

\method{QuickPlot}{abf}(abf, pos = NULL, intv = NULL, cursor = NULL,
  time_unit = NULL, colour = FALSE, title = NULL,
  legend_title = NULL, zero_intercept = TRUE, zero_axes = TRUE,
  line_size = 0.5)

\method{QuickPlot}{data.frame}(df, colour = FALSE, title = NULL,
  legend_title = NULL, zero_intercept = TRUE, zero_axes = TRUE,
  line_size = 0.5, marker_size = line_size * 4,
  err_bar_size = line_size/1.75, err_bar_width = marker_size * 1.25)

\method{QuickPlot}{list}(data, pos = NULL, colour = TRUE,
  title = NULL, legend_title = NULL, zero_intercept = TRUE,
  zero_axes = TRUE, line_size = 0.5, marker_size = line_size * 4,
  err_bar_size = line_size/1.75, err_bar_width = marker_size * 1.25)
}
\arguments{
\item{data}{data to plot, QuickPlot will try its best to figure out how to plot.}

\item{...}{not used.}

\item{pos}{positions to sample and calculate values, can be interval or cursor.}

\item{intv}{interval compatible with PlotChannel(), also used to sample and
calculate values.}

\item{cursor}{cursor compatible with PlotChannel(), also used to sample and
calculate values.}

\item{time_unit}{unit of time.}

\item{colour}{whether to plot in colour mode.}

\item{title}{OPTIONAL, title for the plot.}

\item{legend_title}{OPTIONAL, legend title for the plot.}

\item{zero_intercept}{whether to add zero intercepts.}

\item{zero_axes}{whether to force position axes at 0.}

\item{line_size}{size of lines.}

\item{marker_size}{size of markers, not used when plotting abf objects.}

\item{err_bar_size}{size of error bars, not used when plotting abf objects.}

\item{err_bar_width}{width of error bars, not used when plotting abf objects.}
}
\value{
a ggplot object
}
\description{
QuickPlot automatically select proper plotting functions according to the type
of data and arguments provided. You can always try QuickPlot first if you don't
know how to plot or what plotting function to use before consulting manual.
}
\details{
When plotting an abf object, the type of generated plot is determined by this
logic: 1. Whether the abf object has only one channel, if so, a channel plot is
generated. 2. Whether time_unit is given, if so, plot all channels and convert
x axis to time_unit. 3. Whether the abf object has only one episode, if so, a
channel plot of time unit "tick" is generated.
}
