% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuickPlot.R
\name{QuickPlot_IVSummary}
\alias{QuickPlot_IVSummary}
\title{Quick plot an IV Summary table}
\usage{
QuickPlot_IVSummary(df_summary, err_bar_width = 1.5, title = NULL)
}
\arguments{
\item{df_summary}{an IV summary table from IVSummary()}

\item{err_bar_width}{width of the error bar}

\item{title}{OPTIONAL, title of the plot}
}
\value{
a ggplot object.
}
\description{
Quick plot an IV Summary table
}
