% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuickPlot.R
\name{QuickPlotIV}
\alias{QuickPlotIV}
\title{Quick plot I-V curves at given position}
\usage{
QuickPlotIV(abf, pos, colour = FALSE, title = NULL,
  legend_title = NULL, zero_intercept = TRUE, zero_axes = TRUE)
}
\arguments{
\item{abf}{an abf or a list of abf objects.}

\item{pos}{an interval/cursor or a list of intervals/cursors.}

\item{colour}{whether to plot in coloured mode.}

\item{title}{OPTIONAL, title for the plot.}

\item{zero_intercept}{whether to add zero intercepts to the plot.}

\item{zero_axes}{whether to plot axes at 0.}
}
\value{
a ggplot object.
}
\description{
Quick plot I-V curves at given position
}
