% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuickPlot.R
\name{QuickPlotTrace}
\alias{QuickPlotTrace}
\title{Quick plot trace vs. trace curve.}
\usage{
QuickPlotTrace(abf, channelX, episodeX, channelY, episodeY, intv)
}
\arguments{
\item{abf}{an abf object.}

\item{channelX}{channel of X trace.}

\item{episodeX}{episode of X trace.}

\item{channelY}{channel of Y trace.}

\item{episodeY}{channel of Y trace.}

\item{intv}{interval of the traces to be plotted.}
}
\value{
a ggplot object.
}
\description{
Quick plot trace vs. trace curve.
}
