% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_reduction.R
\name{reduce_along}
\alias{reduce_along}
\alias{reduce_lastdim}
\title{Reduction of x along a specific dimension}
\usage{
reduce_along(x, along, reduce_func = NULL, names = NULL, ...)

reduce_lastdim(x, reduce_func = NULL, names = NULL, ...)
}
\arguments{
\item{x}{data to reduce.}

\item{along}{axis/dimension to reduce.}

\item{reduce_func}{a function to perform the reduction.}

\item{names}{set names to results.}

\item{...}{passed to reduce_func.}
}
\value{
a list
}
\description{
mapnd() and reduce_dim() is not true map-reduce since mapnd() does not emit
anything. However, these functions are designed to mimic the map-reduce paradigm
in hope of facilitating a uniform data processing scheme for all objects.
}
\details{
reduce_lastdim() reduces x along last dimension, and it is a bit more efficient
since no permutation is done to x.
}
\examples{
\dontrun{
max_avg_current <- function(abf) {
  abf \%>\%
     samplend(ratio = 1000, func = "mean") \%>\%
     reduce_lastdim(reduce_func = max)
}

plot_iv <- function(abf, intv) {
  p <- abf[MaskIntv(intv),,] \%>\%
          mapnd(func = "mean") \%>\%
          reduce_lastdim(names = c("current", "voltage")) \%>\%
          data.frame() \%>\%
          ggplot(mapping = aes_string("current", "voltage"))
  p + geom_line()
}
}
}
