% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode.R
\name{RemoveEpisodes}
\alias{RemoveEpisodes}
\title{Remove episodes from an abf object, by-ref behaviour.}
\usage{
RemoveEpisodes(abf, episodes)
}
\arguments{
\item{abf}{an abf or a list of abf object.}

\item{episodes}{episodes to remove.}
}
\value{
an abf object with desired episodes marked removed.
}
\description{
Function behaviour: episodes will be removed from ALL channels. Removing
episodes only affect results of [[ extracting and GetAvailEpisodes, since
episodic data is not actually removed but marked "removed"
}
