% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel.R
\name{ReplaceChannel}
\alias{ReplaceChannel}
\title{Replacing channel data, by-ref like behaviour.}
\usage{
ReplaceChannel(abf, channel_data, channel = 1L)
}
\arguments{
\item{abf}{an abf object.}

\item{channel_data}{channel data to replace the original.}

\item{channel}{ADC channel id, 1-based.}
}
\value{
an abf object with the replaced channel.
}
\description{
Replacing channel data, by-ref like behaviour.
}
