% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel.R
\name{RescaleChannel}
\alias{RescaleChannel}
\alias{RsclChan}
\title{Rescale channel unit.}
\usage{
RescaleChannel(abf, channel, scale = c("1", "pico", "nano", "micro",
  "milli", "kilo", "mega", "Giga", "Tera"))

RsclChan(abf, channel = GetAllChannels(abf), scale = c("1", "pico",
  "nano", "micro", "milli", "kilo", "mega", "Giga", "Tera"))
}
\arguments{
\item{abf}{an abf object}

\item{channel}{channel to rescale}

\item{scale}{scale to use}
}
\value{
abf itself
}
\description{
Rescale channel unit.
}
