% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{RmblAbf}
\alias{RmblAbf}
\alias{RemoveBaselineAbf}
\title{Remove baseline of a channel.}
\usage{
RmblAbf(abf, channel, method = c("interval", "episode", "holding"), intv,
  episode, by_epoch = TRUE, dac = GetWaveformEnabledDAC(abf), ...)

RemoveBaselineAbf(abf, channel, method = c("interval", "episode",
  "holding"), intv, episode, by_epoch = TRUE, dac, ...)
}
\arguments{
\item{abf}{an abf object.}

\item{channel}{a channel to perform baseline removal}

\item{method}{method to use. See details.}

\item{intv}{used in interval method. An interval to evaluate baseline level.}

\item{episode}{used in episode method. The episode to evaluate baseline.}

\item{by_epoch}{used in episode method, see \code{\link[abftools:EpisodeBaseline]{EpisodeBaseline()}} for details.}

\item{dac}{used in holding and episode method, waveform dac channel.}

\item{...}{used in episode method, passed to baseline_als, see \code{\link[abftools:baseline_als]{baseline_als()}} for details.}
}
\value{
an abf object.
}
\description{
Remove baseline of a channel.
}
\details{
Avaiable methods for baseline removal are holding, interval and episode.
The interval method evaluate mean baseline level from the given episode between intv,
then subtract the value to every episode.
The episode method calculates baseline by calling \code{\link[abftools:baseline_als]{baseline_als()}}.
The holding method simply read holding level from DAC setting and subtract the
value to every episode. This is usually desirable only for waveform channel.
}
