% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{SampleAbf}
\alias{SampleAbf}
\title{Sample abf object to reduce data points, by-ref like behaviour.}
\usage{
SampleAbf(abf, sample_ratio, sample_func = NULL, ...)
}
\arguments{
\item{abf}{an abf object.}

\item{sample_ratio}{the sampling ratio. See melt.abf for more details.}

\item{sample_func}{a sampling function applied to sampled points.}

\item{...}{further arguments passed to sampling_colFunc}
}
\value{
the sampled abf itself
}
\description{
Sample abf object to reduce data points, by-ref like behaviour.
}
