% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_base.R
\name{samplend}
\alias{samplend}
\alias{samplend_col}
\title{Sample an nd-array evenly by ratio.}
\usage{
samplend(x, ratio = 1L, func = NULL, along = 1L, ...)

samplend_col(x, ratio = 1L, colFunc = NULL, along = 1L, ...)
}
\arguments{
\item{x}{an nd-array.}

\item{ratio}{a sampling ratio.}

\item{func}{a sampling function.}

\item{along}{the axis/dimension to sample along.}

\item{...}{passed to func/colFunc.}

\item{colFunc}{a column function.}

\item{func_name}{a optimised function name.}
}
\value{
a sampled nd-array.
}
\description{
Like mapnd() and mapnd_col(), for performance consideration, samplend_col is
strongly encouraged is a sampling function is used.
}
