% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{scale_unit}
\alias{scale_unit}
\title{Basic scaling of units.}
\usage{
scale_unit(x, unit, scale = c("1", "pico", "nano", "micro", "milli",
  "kilo", "mega", "Giga", "Tera"))
}
\arguments{
\item{x}{a numeric vector.}

\item{unit}{unit of x.}

\item{scale}{new scale of prefix}
}
\value{
a numeric vector
}
\description{
Scale x of unit to a new unit with different prefix
}
