% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_indexing.R
\name{SelectSample}
\alias{SelectSample}
\title{Select samples conditionally}
\usage{
SelectSample(sample_index, ...)
}
\arguments{
\item{sample_index}{a data frame like object with descriptions of samples}

\item{...}{a list of conditions composed of name = value pairs.}
}
\value{
selected rows of sample_index
}
\description{
This function provides a naïve yet simple way to select desired samples provided
in sample_index. A sample_index should have named columns which descpribes
properties of your samples, any unnamed columns or conditions will be ignored.
Be noticed that all names are case-sensitive.
}
