% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{sem_abf}
\alias{sem_abf}
\title{Calculate standard error of an abf object.}
\usage{
sem_abf(abf, intv = NULL, ret.df = FALSE, use_chan_name = FALSE,
  na.rm = TRUE)
}
\arguments{
\item{abf}{an abf object.}

\item{intv}{OPTIONAL, an interval to calculate.}

\item{ret.df}{whether to return a data.frame object, if set to FALSE a matrix is returned instead.}

\item{use_chan_name}{whether to use channel names as colnames, set to FALSE to use descriptive names instead.}

\item{na.rm}{whether to remove NA values.}
}
\value{
A matrix/data.frame.
}
\description{
Returns a matrix/data.frame containing episodic SEM of each channel.
}
