% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R
\name{SetChanDesc}
\alias{SetChanDesc}
\title{Set channel description.}
\usage{
SetChanDesc(abf, desc, channel = 1L)
}
\arguments{
\item{abf}{an abf object.}

\item{desc}{a new description for the channel.}

\item{channel}{channel id.}
}
\value{
a abf object with new ChannelDesc set.
}
\description{
Set channel description.
}
