% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R
\name{SetChannelName}
\alias{SetChannelName}
\title{Set channel name, by-ref behaviour.}
\usage{
SetChannelName(abf, name, channel = 1L)
}
\arguments{
\item{abf}{an abf object.}

\item{name}{a new name}

\item{channel}{channel id.}
}
\value{
a abf object with new name set, invisibly.
}
\description{
Set channel name, by-ref behaviour.
}
