% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SetIntv}
\alias{SetIntv}
\title{Set intervals, by-ref behaviour}
\usage{
SetIntv(intv, startPos, endPos)
}
\arguments{
\item{intv}{an interval.}

\item{startPos}{starting position of the new interval.}

\item{endPos}{ending position of the new interval.}
}
\value{
an interval.
}
\description{
Set intervals, by-ref behaviour
}
