% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{SpIVSummary}
\alias{SpIVSummary}
\alias{SpIGVSummary}
\title{Report I-V relationship of a list of abf objects, with specific current/conductance.}
\usage{
SpIVSummary(abf, intv = NULL, conductance = FALSE, memprops = NULL,
  current_channel = GetFirstCurrentChan(abf),
  voltage_channel = GetFirstVoltageChan(abf), group = NA,
  unit = FALSE, ...)

SpIGVSummary(abf, intv = NULL, memprops = NULL,
  current_channel = GetFirstCurrentChan(abf),
  voltage_channel = GetFirstVoltageChan(abf), group = NA,
  unit = FALSE, ...)
}
\arguments{
\item{abf}{a list of abf objects.}

\item{intv}{OPTIONAL, a list of intervals.}

\item{conductance}{Only used in SpIVSummary(), if TRUE, same as calling SpIGVSummary()}

\item{memprops}{membrane properties, if NULL memprops will be calculated from StepMemtestSummary().}

\item{current_channel}{current channel id, 1-based.}

\item{voltage_channel}{voltage channel id, 1-based.}

\item{group}{OPTIONAL, add a "Group" column to the returned data.frame.}

\item{unit}{whether to add unit columns to the returned data.frame.}

\item{...}{passed to StepmemtestSummary(), \code{\link[abftools:StepmemtestSummary]{help}} for details.}
}
\value{
a data.frame
}
\description{
Report I-V relationship of a list of abf objects, with specific current/conductance.
}
