% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacitance.R
\name{StepMemtestAbf}
\alias{StepMemtestAbf}
\title{Calculate membrane properties from a step pulse membrane test.}
\usage{
StepMemtestAbf(abf, intv = NULL, episode = NULL,
  epoch = FindMemtestEpoch(abf, dac = dac, type = "step"),
  dac = GetWaveformEnabledDAC(abf),
  current_channel = GetFirstCurrentChan(abf), max_iter = 30L,
  simplify = TRUE)
}
\arguments{
\item{abf}{an abf object.}

\item{intv}{interval of the CHARGING period.}

\item{episode}{episode of the memtest.}

\item{epoch}{epoch of the memtest.}

\item{dac}{DAC channel of Vcmd}

\item{current_channel}{current channel.}

\item{max_iter}{maximum iterations.}

\item{simplify}{whether to return a flat simplified named vector.}
}
\value{
a list of membrane properties, see Details.
}
\description{
Calculate membrane properties from a step pulse membrane test.
}
\details{
This function calculates membreane properties given a proper step
pulse is applied and recorded in the abf object. Usually all arguments can
be determined automatically. By default episode is determined by
\code{\link[abftools:FindStepEpisode]{FindStepEpisode()}}, see help for more
details.

The returned list contains membrane properties calculated, which are:

Rtot: total resistance (in Ohms)

Rs: series resistance (in Ohms)

Rm: membrane resistance (in Ohms)

Cm: membrane capacitance (in Farad)

coefs: Fitted coefficients for debug purpose.

coefs$Is: Steady current (in A)

coefs$I0: Max current (in A)

coefs$lrc: log rate constant (arbitrary unit, logarithm of second)

coefs$tao: RC time constant (in s)

coefs$fitted: fitted model.
}
