% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacitance.R
\name{StepMemtestSummary}
\alias{StepMemtestSummary}
\title{Report membrane properties by StepMemtestAbf() for a list of abf objects.}
\usage{
StepMemtestSummary(abf, intv_charge = NULL, episode = NULL,
  epoch = NULL, dac = NULL, current_channel = NULL, max_iter = 10L)
}
\arguments{
\item{abf}{a list of abf object.}

\item{intv_charge}{interval of the charging period.}

\item{episode}{episodes to calculate membrane properties.}

\item{epoch}{epoch of the step pulse.}

\item{dac}{DAC channel of Vcmd.}

\item{current_channel}{current channel.}

\item{max_iter}{max iterations.}
}
\value{
a data.frame
}
\description{
Report membrane properties by StepMemtestAbf() for a list of abf objects.
}
