% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abf2_episodic_s3fun.R
\name{[.ABF2.Episodic}
\alias{[.ABF2.Episodic}
\title{Extract or Replace channel data of an episodic abf2 object}
\usage{
\method{[}{ABF2.Episodic}(x, i, j, channel = 1)
}
\arguments{
\item{x}{episodic abf2 object}

\item{i}{rows (data points) to extract or replace}

\item{j}{columns (episodes) to extract or replace}

\item{channel}{channel to extract or replace}
}
\description{
Extract or Replace channel data of an episodic abf2 object
}
