% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abf2_episodic_s3fun.R
\name{[[<-.ABF2.Episodic}
\alias{[[<-.ABF2.Episodic}
\title{Unsupported op [[<-.ABF2.Episodic}
\usage{
\method{[[}{ABF2.Episodic}(x, channel) <- value
}
\arguments{
\item{value}{}
}
\description{
Unsupported op [[<-.ABF2.Episodic
}
