% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abf2_episodic_s3fun.R
\name{[<-.ABF2.Episodic}
\alias{[<-.ABF2.Episodic}
\title{[<-.ABF2.Episodic}
\usage{
\method{[}{ABF2.Episodic}(x, i, j, channel = 1) <- value
}
\arguments{
\item{value}{}
}
\description{
[<-.ABF2.Episodic
}
