% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{wrap}
\alias{wrap}
\title{Convenient alias of WrapMappingFunc()}
\usage{
wrap(...)
}
\arguments{
\item{...}{See WrapMappingFunc for more details.}
}
\value{
a function of f(abf, intv)
}
\description{
Convenient alias of WrapMappingFunc()
}
