% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{WrapMappingFuncAlong}
\alias{WrapMappingFuncAlong}
\alias{wrap_along}
\title{Wrap a mapping function along specific axis to batch process abf data.}
\usage{
WrapMappingFuncAlong(map_func, along = c("time", "episode", "channel"),
  pack_args = FALSE, abf_id_func = NULL,
  epi_id_func = DefaultEpiLabel, chan_id_func = DefaultChanLabel,
  time_unit = "tick", ret.df = TRUE, ...)

wrap_along(..., time_unit = "ms", ret.df = TRUE)
}
\arguments{
\item{map_func}{a mapping function.}

\item{along}{the axis to process along. Can be "time", "episode" or "channel".}

\item{pack_args}{wheter to pack arguments for map_func, see PackArgs() for more details.}

\item{abf_id_func}{OPTIONAL, a function accepts an abf object and returns an identifier of the objects.}

\item{epi_id_func}{OPTIONAL, a function accepts an abf object and returns a vector of identifiers of all episodes.}

\item{chan_id_func}{OPTIONAL, a function accepts an abf object and returns a vectors of identifiers of all channels.}

\item{time_unit}{convert time unit of the return time axis.}

\item{ret.df}{wheter to return a data.frame, a matrix is returned instead if set to FALSE.}

\item{...}{further arguments passed to map_func.}
}
\value{
a function
}
\description{
The wrapped function has a signature of f(abf, intv = NULL, episode, channel)
see details for more details.
}
\details{
Arguments of the wrapped function f(abf, intv = NULL, episode, channel):

abf   a abf object

intv   OPTIONAL, a time interval to calcaulte. If NULL, the whole timespan is used.

episode   OPTIONAL, a vector of episodes to calculate. If missing, all episodes are used.

channel   OPTIONAL, a vector of channels to calculate. If missing, all channels are used.
}
