% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{ZeroAxes}
\alias{ZeroAxes}
\title{Shift the axes of a ggplot object to 0}
\usage{
ZeroAxes(xlimit, ylimit, xlabel, ylabel, xticks = 5, yticks = 5,
  textsize = 5)
}
\arguments{
\item{xlimit}{limit of x values}

\item{ylimit}{limit of y values}

\item{xlabel}{label of x axis}

\item{ylabel}{label of y axis}

\item{xticks}{number of ticks of x axis}

\item{yticks}{number of ticks of y axis}

\item{textsize}{size of axis label size}
}
\value{
a list of ggplot object
}
\description{
Shift the axes of a ggplot object to 0
}
