% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abf2_load.R
\name{abf2_load}
\alias{abf2_load}
\title{abf2_load}
\usage{
abf2_load(filename, abf_title = NULL)
}
\arguments{
\item{abf_title}{}
}
\value{
ABF2.Episodic, ABF2.Gapfree depending on file type.
}
\description{
abf2_load
}
