% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abf2_load.R
\name{abf2_load}
\alias{abf2_load}
\title{abf2_load}
\usage{
abf2_load(filename, abf_title = NULL)
}
\arguments{
\item{filename}{file path to the abf2 file.}

\item{abf_title}{a title assigned to the loaded file, default is filename.}
}
\value{
an abf object.
}
\description{
abf2_load
}
