% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abf2_load.R
\name{abf2_loadlist}
\alias{abf2_loadlist}
\title{abf2_loadlist}
\usage{
abf2_loadlist(filelist, folder = NULL, attach_ext = TRUE,
  titlelist = NULL, short_desc = TRUE)
}
\arguments{
\item{filelist}{a list of file name or a data.frame like object with a column named FileName/filename.}

\item{folder}{OPTIONAL, the path to the folder of the files, if filelist contains full path, leave this to empty.}

\item{attach_ext}{automatically add ".abf" extension to filelist if not present.}

\item{titlelist}{OPTIONAL, a list of titles, a single title to be assigned to all loaded files.}

\item{short_desc}{whether to use short channel description.}
}
\value{
a list of abf objects.
}
\description{
abf2_loadlist
}
