% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_content.R
\name{abf_plot_cd}
\alias{abf_plot_cd}
\title{Get an abf channel domain ggplot object}
\usage{
abf_plot_cd(abf, intv = NULL, channel = c(2L, 1L), map_func,
  pack_args = FALSE, ..., colour = FALSE)
}
\description{
Get an abf channel domain ggplot object
}
