% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_content.R
\name{abf_plot_td}
\alias{abf_plot_td}
\title{Get an abf time domain ggplot object}
\usage{
abf_plot_td(abf, intv = NULL, channel = 1L, sample_ratio = 1L,
  sample_func = NULL, sample_colFunc = NULL, ..., time_unit = "tick",
  colour = FALSE)
}
\description{
Get an abf time domain ggplot object
}
