% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abf2_episodic_tricks.R
\name{AllSamples_IVSummary}
\alias{AllSamples_IVSummary}
\title{AllSamples_IVSummary calculates average current of a list of abf objects, within given intervals}
\usage{
AllSamples_IVSummary(abf_list, intv_list, current_channel = 0,
  voltage_channel = 0)
}
\arguments{
\item{voltage_channel}{}
}
\description{
AllSamples_IVSummary calculates average current of a list of abf objects, within given intervals
}
