% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_base.R
\name{apply_unsafe}
\alias{apply_unsafe}
\title{An unsafe apply()}
\usage{
apply_unsafe(x_, along_, func_, keep_dim_ = FALSE, ...)
}
\arguments{
\item{x_}{an array.}

\item{along_}{dimension to apply along.}

\item{func_}{a function to apply.}

\item{keep_dim_}{keep dim order}

\item{...}{passed to func_}
}
\value{
an array.
}
\description{
Basically apply() without validating anything.
}
