% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel.R
\name{AtchChan}
\alias{AtchChan}
\title{Attach a new channel to an abf object.}
\usage{
AtchChan(abf, channel_data, channel_name, channel_unit,
  channel_desc = channel_name)
}
\arguments{
\item{abf}{an abf object.}

\item{channel_data}{data of the new channel.}

\item{channel_name}{name of the new channel.}

\item{channel_unit}{unit of the new channel.}

\item{channel_desc}{description of the new channel.}
}
\value{
an abf object with the attached channel.
}
\description{
The attached channel_data's dimensions must match original dimensions,
regardless of removed episodes due to the mechanism of RemoveEpisodes (they
are only marked removed and excluded when extracting using [[).
}
