% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveform.R
\name{AttachWaveform}
\alias{AttachWaveform}
\title{Attach a waveform channel to an abf object, by-ref behaviour.}
\usage{
AttachWaveform(abf)
}
\arguments{
\item{abf}{an abf object.}
}
\value{
an abf object with a new waveform channel attached.
}
\description{
Attach a waveform channel to an abf object, by-ref behaviour.
}
