% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{AverageAbf}
\alias{AverageAbf}
\title{Average a list of abf objects.}
\usage{
AverageAbf(abf_list, w = NULL)
}
\arguments{
\item{abf_list}{a list of abf objects.}

\item{w}{OPTIONAL, a vector of weights for weighted average.}
}
\value{
an averaged abf object, of which the protocol settings follow first element in abf_list.
}
\description{
Average a list of abf objects.
}
