% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{BlankAbf}
\alias{BlankAbf}
\title{Blank current channel of abf objects.}
\usage{
BlankAbf(abf, ref_data, ref_intv = NULL, current_channel = NULL)
}
\arguments{
\item{abf}{an abf object or a list of abf objects.}

\item{ref_data}{reference current data to blank or abf objects as reference.}

\item{ref_intv}{OPTIONAL, if ref_data is abf objects, a time interval to
calculate mean current.}

\item{current_channel}{OPTIONAL, the current channel to apply blanking,
if not given, the first current channel is used.}
}
\value{
an abf object or a list of abf objects
}
\description{
Blank current channel of abf objects.
}
