% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autointv.R
\name{CmpWaveform}
\alias{CmpWaveform}
\title{Compare epoch waveform setting to a channel data.}
\usage{
CmpWaveform(abf, episode = GetAllEpisodes(abf), channel, epoch,
  delta = NULL, min_win = NULL, max_win = NULL)
}
\arguments{
\item{abf}{an abf object.}

\item{episode}{the episodes to compare.}

\item{channel}{the channel to compare, channel id is 1-based.}

\item{epoch}{the epoch to compare.}

\item{delta}{allowed max deviation.}

\item{min_win}{OPTIONAL, minimum interval window size of the result.}

\item{max_win}{OPTIONAL, maximum interval window size of the result.}
}
\value{
a list of intervals of which pass comparison.
}
\description{
Compare epoch waveform setting to a channel data.
}
