% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colSems}
\alias{colSems}
\title{Calculate column standard error of measurements}
\usage{
colSems(df, na.rm = FALSE)
}
\arguments{
\item{df}{a data frame of data frame like 2d data}

\item{na.rm}{remove NAs}
}
\value{
calculated column sem
}
\description{
Calculate column standard error of measurements
}
