% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodic.R
\name{EpisodicIntervalMeans}
\alias{EpisodicIntervalMeans}
\title{EpisodicIntervalMeans calculates interval means of a list of abf objects}
\usage{
EpisodicIntervalMeans(abf_list, intv_list, channel = 1, na.rm = TRUE)
}
\arguments{
\item{na.rm}{}
}
\description{
EpisodicIntervalMeans calculates interval means of a list of abf objects
}
