% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodic.R
\name{EpisodicIntervalSds}
\alias{EpisodicIntervalSds}
\title{EpisodicIntervalSds calculates interval standard deviations of a list of abf objects}
\usage{
EpisodicIntervalSds(abf_list, intv_list, channel = 1, na.rm = TRUE)
}
\arguments{
\item{na.rm}{}
}
\description{
EpisodicIntervalSds calculates interval standard deviations of a list of abf objects
}
