% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_indexing.R
\name{ExcludeSample}
\alias{ExcludeSample}
\title{Exclude samples conditionally}
\usage{
ExcludeSample(sample_index, ...)
}
\arguments{
\item{sample_index}{a data frame like object with descriptions of samples}

\item{...}{a list of conditions composed of name = value pairs.}
}
\value{
remaining rows of sample_index
}
\description{
This function provides a naïve yet simple way to exclude unwanted samples
provided in sample_index. A sample_index should have named columns which
descpribe properties of your samples, any unnamed columns or conditions will
be ignored. Be noticed that all names are case-sensitive.
}
