% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autointv.R
\name{FindAllSamplingInterval}
\alias{FindAllSamplingInterval}
\title{Call FindSamplingInterval for a list of abf objects.}
\usage{
FindAllSamplingInterval(abf_list, ...)
}
\arguments{
\item{abf_list}{a list of abf objects.}

\item{...}{}
}
\value{
a list of found intervals.
}
\description{
Call FindSamplingInterval for a list of abf objects.
}
