% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autointv.R
\name{FindSamplingInterval}
\alias{FindSamplingInterval}
\alias{FindAllSamplingInterval}
\title{FindSamplingInterval finds a stable interval for sampling current and voltage data of an abf object.}
\usage{
FindSamplingInterval(abf, epoch = "B",
  current_channel = GetFirstCurrentChan(abf),
  voltage_channel = GetFirstVoltageChan(abf),
  target_interval_size = NULL, allowed_voltage_delta = NULL,
  backward_search = TRUE, noisy_data = FALSE)

FindAllSamplingInterval(abf_list, ...)
}
\arguments{
\item{abf}{an abf object.}

\item{epoch}{OPTIONAL, the epoch to search, defaults to B (second epoch).}

\item{current_channel}{OPTIONAL, channel id for current data, channel id is 1-based.}

\item{voltage_channel}{OPTIONAL, channel id for voltage data, channel id is 1-based.}

\item{target_interval_size}{OPTIONAL, target size in **points** of the sampling interval. Default is 20ms of samples.}

\item{allowed_voltage_delta}{OPTIONAL, allowed max deviation of voltage.}

\item{backward_search}{OPTIONAL, perform search along backward direction.}

\item{noisy_data}{OPTIONAL, set to TURE if data is noisy, may improve position of the predicted sampling interval.}
}
\value{
a named vector of 3 numeric: interval start position, end position, length
}
\description{
FindSamplingInterval is a convenient alternative to CmpWaveform, best suited
to locate sampling intervals when using step voltage cmd waveform. Some common
assumptions are made:
1. Waveform Cmd is outputing voltage command.
2. The function is looking for most stable voltage AND current channel.
3. The function prefers stability to the size of interval.
4. Intervals closer to the end of the epoch are preferred.
}
