% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abf2_episodic_tricks.R
\name{FindSamplingInterval}
\alias{FindSamplingInterval}
\title{FindSamplingInterval finds a stable interval for sampling current and voltage data of an abf object}
\usage{
FindSamplingInterval(abf, current_chan_id = 0, voltage_chan_id = 0,
  min_sampling_size = 0, max_sampling_size = 0,
  allowed_voltage_delta = 0, epoch_name = "auto",
  backward_search = TRUE, strict_comp = TRUE)
}
\arguments{
\item{abf}{an abf object}

\item{current_chan_id}{OPTIONAL, channel id for current data}

\item{voltage_chan_id}{OPTIONAL, channel id for voltage data}

\item{min_sampling_size}{OPTIONAL, min size in points of a sampling interval}

\item{allowed_voltage_delta}{OPTIONAL, allowed max deviation of voltage W.R.T epoch per DAC setting}

\item{epoch_name}{}

\item{backward_search}{OPTIONAL, perform search along backward direction}

\item{strict_comp}{OPTIONAL, use strict compare mode, if set to FALSE, a fuzzy "mostly" compare is used instead}
}
\value{
a named vector of 3 numeric: interval start position, end position, length
}
\description{
FindSamplingInterval finds a stable interval for sampling current and voltage data of an abf object
}
