% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autointv.R
\name{FindSamplingInterval}
\alias{FindSamplingInterval}
\alias{FindAllSamplingInterval}
\title{FindSamplingInterval finds a stable interval for sampling current and voltage data of an abf object.}
\usage{
FindSamplingInterval(abf, epoch = NULL,
  dac = GetWaveformEnabledDAC(abf),
  current_channel = GetFirstCurrentChan(abf),
  voltage_channel = GetFirstVoltageChan(abf),
  target_interval_size = NULL, allowed_voltage_delta = NULL,
  noisy_opt = FALSE, lp_freq = 75, lp_order = 1L)

FindAllSamplingInterval(abf_list, ...)
}
\arguments{
\item{abf}{an abf object.}

\item{epoch}{the epoch to search, defaults to first multi-step epoch.}

\item{dac}{the dac channel to evaluate waveform.}

\item{current_channel}{channel id for current data, channel id is 1-based.}

\item{voltage_channel}{channel id for voltage data, channel id is 1-based.}

\item{target_interval_size}{OPTIONAL, target size in **points** of the sampling interval.
Default is 2x lowpass window.}

\item{allowed_voltage_delta}{OPTIONAL, allowed max deviation of voltage.}

\item{noisy_opt}{enable optimisation for noisy data by applying a lowpass filter to current channel.}

\item{lp_freq}{frequency of low-pass filter.}

\item{lp_order}{order of low-pass filter.}
}
\value{
a named vector of 3 numeric: interval start position, end position, length
}
\description{
FindSamplingInterval is a convenient alternative to CmpWaveform, best suited
to locate sampling intervals when using step voltage cmd waveform. Some common
assumptions are made:
1. Waveform Cmd is outputing voltage command.
2. Waveform epochs are aligned, i.e. epoch duration incr is 0.
3. The function is looking for most stable voltage AND current channel.
4. The function prefers stability to the size of interval.
5. Intervals closer to the end of the epoch are preferred.
}
