% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autointv.R
\name{FindSamplingInterval}
\alias{FindSamplingInterval}
\title{FindSamplingInterval finds a stable interval for sampling current and voltage data of an abf object}
\usage{
FindSamplingInterval(abf, current_channel = 0, voltage_channel = 0,
  min_sampling_size = 0, allowed_voltage_delta = 0, epoch_name = "B",
  backward_search = TRUE)
}
\arguments{
\item{abf}{an abf object}

\item{current_channel}{OPTIONAL, channel id for current data}

\item{voltage_channel}{OPTIONAL, channel id for voltage data}

\item{min_sampling_size}{OPTIONAL, min size in points of a sampling interval}

\item{allowed_voltage_delta}{OPTIONAL, allowed max deviation of voltage W.R.T epoch per DAC setting}

\item{epoch_name}{OPTIONAL, the epoch to search, defaults to B (second epoch)}

\item{backward_search}{OPTIONAL, perform search along backward direction}
}
\value{
a named vector of 3 numeric: interval start position, end position, length
}
\description{
FindSamplingInterval finds a stable interval for sampling current and voltage data of an abf object
}
