% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autointv.R
\name{FindSamplingInterval}
\alias{FindSamplingInterval}
\title{FindSamplingInterval finds a stable interval for sampling current and voltage data of an abf object.}
\usage{
FindSamplingInterval(abf, current_channel, voltage_channel,
  min_sampling_size, max_sampling_size, allowed_voltage_delta,
  epoch_name = "B", backward_search = TRUE)
}
\arguments{
\item{abf}{an abf object.}

\item{current_channel}{OPTIONAL, channel id for current data, channel id is 1-based.}

\item{voltage_channel}{OPTIONAL, channel id for voltage data, channel id is 1-based.}

\item{min_sampling_size}{OPTIONAL, min size in points of a sampling interval.}

\item{max_sampling_size}{OPTIONAL, max size in points of a sampling interval.}

\item{allowed_voltage_delta}{OPTIONAL, allowed max deviation of voltage.}

\item{epoch_name}{OPTIONAL, the epoch to search, defaults to B (second epoch).}

\item{backward_search}{OPTIONAL, perform search along backward direction.}
}
\value{
a named vector of 3 numeric: interval start position, end position, length
}
\description{
FindSamplingInterval is a convenient alternative to CmpWaveform, best suited
to locate sampling intervals when using step voltage cmd waveform. Some common
assumptions are made:
1. Waveform Cmd is outputing voltage command.
2. The function is looking for most stable voltage AND current channel.
3. The function prefers stability to the size of interval.
4. Intervals closer to the end of the epoch are preferred.
}
