% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aux.R
\name{FreqToTick}
\alias{FreqToTick}
\title{Calculate time span of a frequency}
\usage{
FreqToTick(freq, sampling_rate)
}
\arguments{
\item{freq}{a frequency in Hz.}

\item{sampling_rate}{sampling rate (in µs), or an abf object that sampling rate is extracted from.}
}
\value{
an integer of **point**/**tick** count.
}
\description{
Calculate time span of a frequency
}
