% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R
\name{GetChannelName}
\alias{GetChannelName}
\title{Get names of channels.}
\usage{
GetChannelName(abf)
}
\arguments{
\item{abf}{an abf object.}
}
\value{
a character vector of channel names.
}
\description{
Get names of channels.
}
