% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{GetEpiLabel}
\alias{GetEpiLabel}
\title{Get label for every episode of an abf object}
\usage{
GetEpiLabel(abf, style)
}
\arguments{
\item{abf}{an abf object}

\item{style}{a format string}
}
\value{
a vector of characters.
}
\description{
Get label for every episode of an abf object
}
