% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R
\name{GetPointsPerEpisode}
\alias{GetPointsPerEpisode}
\alias{nPts}
\title{Get number of recorded points per episode/sweep.}
\usage{
GetPointsPerEpisode(abf)

nPts(abf)
}
\arguments{
\item{abf}{an abf object.}
}
\value{
number of recorded points per episode/sweep.
}
\description{
Get number of recorded points per episode/sweep.
}
