% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R
\name{GetSamplingIntv}
\alias{GetSamplingIntv}
\alias{GetSamplingRate}
\title{Get sampling interval in us.}
\usage{
GetSamplingIntv(abf)

GetSamplingRate(abf)
}
\arguments{
\item{abf}{an abf object.}
}
\value{
the sampling interval in unit us.
}
\description{
Get sampling interval in us.
}
