% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_attr.R
\name{GetVoltageChan}
\alias{GetVoltageChan}
\title{Get voltage channel id.}
\usage{
GetVoltageChan(abf)
}
\arguments{
\item{abf}{an abf object or a list of abf objects.}
}
\value{
a vector of voltage channel id.
}
\description{
If abf is a list of abf objects, returns common voltage channel id.
}
