% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveform.R
\name{GetWaveform}
\alias{GetWaveform}
\title{Simulate waveform.}
\usage{
GetWaveform(abf, episodes = NULL, wf_dac_id = NULL)
}
\arguments{
\item{abf}{an abf object.}

\item{episodes}{the episodes to simulate.}

\item{wf_dac_id}{waveform DAC channel, 1-based.}
}
\value{
channel data of the simulated waveform.
}
\description{
Simulate waveform.
}
