% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{IVSummary}
\alias{IVSummary}
\title{IVSummary calculates average current of a list of abf objects, within given intervals}
\usage{
IVSummary(abf_list, intv_list, current_channel, voltage_channel)
}
\arguments{
\item{abf_list}{a list of abf objects.}

\item{intv_list}{OPTIONAL, a list of intervals.}

\item{current_channel}{current channel id, 1-based.}

\item{voltage_channel}{voltage channel id, 1-based.}
}
\value{
a data frame containing calculated Voltage, SEM Voltage, Current, SEM Current columns.
}
\description{
IVSummary calculates average current of a list of abf objects, within given intervals
}
