% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{IVSummary}
\alias{IVSummary}
\alias{IGVSummary}
\title{Report I-V relationship of a list of abf objects.}
\usage{
IVSummary(abf, intv = NULL, conductance = FALSE,
  current_channel = GetFirstCurrentChan(abf),
  voltage_channel = GetFirstVoltageChan(abf), group = NA,
  unit = FALSE)

IGVSummary(abf, intv = NULL,
  current_channel = GetFirstCurrentChan(abf),
  voltage_channel = GetFirstVoltageChan(abf), group = NA,
  unit = FALSE)
}
\arguments{
\item{abf}{a list of abf objects.}

\item{intv}{OPTIONAL, a list of intervals.}

\item{conductance}{Only used in IVSummary(), if TRUE, same as calling IGVSummary()}

\item{current_channel}{current channel id, 1-based.}

\item{voltage_channel}{voltage channel id, 1-based.}

\item{group}{OPTIONAL, add a "Group" column to the returned data.frame.}

\item{unit}{whether to add unit columns to the returned data.frame.}
}
\value{
a data.frame
}
\description{
Report I-V relationship of a list of abf objects.
}
\details{
This function calculates episodic mean values of current and voltage
channels and thier corresponding SEMs given a list of abf objects.
IGVSummary() also calulates conductance. However, while averaging conductances
gives correct results, SEM of conducetance may not properly represent the actual
errors. Though also reported in the returned data.frame, it is not suggested
to plot SEM Conductance since it can be mis-interpreted in some cases.
}
