% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{mapnd}
\alias{mapnd}
\title{Mapping function to an nd-array along specific axis.}
\usage{
mapnd(x, f, along = 1L, pack_args = FALSE, ...)
}
\arguments{
\item{x}{an nd-array.}

\item{f}{a function to map.}

\item{along}{the axis to map the function to.}

\item{pack_args}{whether to pack arguments for f.}

\item{...}{other argumetns passed to f.}
}
\value{
an array (dimension depending on returned dimension of f)
}
\description{
Mapping function to an nd-array along specific axis.
}
