% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode.R
\name{MaskEpisodes}
\alias{MaskEpisodes}
\title{Mask episodes with a value, by-ref behaviour.}
\usage{
MaskEpisodes(abf, channel, episodes, value)
}
\arguments{
\item{abf}{an abf or a list of abf object.}

\item{channel}{channel id, 1-based.}

\item{episodes}{the episodes to mask.}

\item{value}{the value assigned to the episodes.}
}
\value{
an abf object with desired episodes masked with value.
}
\description{
Mask episodes with a value, by-ref behaviour.
}
