% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{mean.abf}
\alias{mean.abf}
\title{Calculate mean values of an abf object.}
\usage{
\method{mean}{abf}(abf, intv = NULL, ret.df = TRUE,
  desc_colnames = TRUE, na.rm = TRUE, ...)
}
\arguments{
\item{abf}{an abf object.}

\item{intv}{OPTIONAL, an interval to mean over.}

\item{ret.df}{whether to return a data.frame object, if set to FALSE a matrix is returned instead.}

\item{desc_colnames}{whether to use descriptive colnames, set to FALSE to use original channel names.}

\item{na.rm}{whether to remove NA values.}

\item{...}{passed to arithmetic mean.}
}
\value{
A data.frame object.
}
\description{
Returns a data.frame object containing episodic mean values of each channel.
}
