% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{mean.abf}
\alias{mean.abf}
\title{Calculates mean values of an abf object.}
\usage{
\method{mean}{abf}(abf, intv, desc_colnames = TRUE, ...)
}
\arguments{
\item{abf}{an abf object.}

\item{intv}{OPTIONAL, an interval to mean over.}

\item{desc_colnames}{Whether to use descriptive colnames, set to FALSE to use original channel names.}

\item{...}{Passed to arithmetic mean.}
}
\value{
A data.frame object
}
\description{
Returns a data.frame object containing episodic mean values of each channel.
}
