% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{melt.abf}
\alias{melt.abf}
\title{Melt channel data of an abf object into a data frame by sampling ratio.}
\usage{
\method{melt}{abf}(abf, channel = 1L, sampling_ratio = 1L,
  sampling_func = NULL, time_unit = "tick")
}
\arguments{
\item{abf}{an abf object}

\item{channel}{the channel to melt, default to 1}

\item{sampling_ratio}{sampling ratio, default to 1 so that no sampling at all}

\item{sampling_func}{a function, default to null so no processing sampled points}

\item{time_unit}{time unit of the sampled data, can be tick, us, ms, s, min or hr}
}
\value{
a melted data frame.
}
\description{
For easier melting abf channel data, sampling time (by given time unit) is
selected as id.vars, variable.name is set to Episode and value.name is
determined by your selected channel, which is extracted from ChannelDesc attr
of the abf object.
}
\details{
You can also provide sampling_ratio and sampling_func to reduce data points
of the molten data frame. A sampling function should accept a matrix as input
and apply column-wise ops to the sampled data and return a numeric (e.g. colMeans),
or be left NULL so no sampling ops is applied at all.
}
