% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{MeltAbf}
\alias{MeltAbf}
\title{Melt an abf object.}
\usage{
MeltAbf(abf, intv = NULL, channel = 1L, sampling_ratio = 1L,
  sampling_func = NULL, sampling_colFunc = NULL,
  time_unit = c("tick", "us", "ms", "s", "min", "hr"), ...,
  value.name = NULL)
}
\arguments{
\item{abf}{an abf object.}

\item{intv}{a time interval to melt.}

\item{channel}{channels to melt.}

\item{sampling_ratio}{a sampling ratio if need to reduce data points.}

\item{sampling_func}{a sampling function if sampling_ratio > 1.}

\item{sampling_colFunc}{a column vectorised sampling function, use instead of
sampling_func for better performance.}

\item{time_unit}{a time unit.}

\item{...}{passed to sampling_func or sampling_colFunc.}

\item{value.name}{a character vector to identity channel columns.}
}
\value{
a data.frame
}
\description{
Melt an abf object into a data.frame grouped by episodes. A time column with
unit provided by time_unit is added to encode time dimension of the orignal
abf object.
}
\details{
For performance consideration, using a column function by sampling_colFunc is
strongly encouraged.
}
