% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{MeltAbf}
\alias{MeltAbf}
\title{Melt an abf object.}
\usage{
MeltAbf(abf, intv = NULL, channel = 1L, sample_ratio = 1L,
  sample_func = NULL, sample_colFunc = NULL, time_unit = c("tick",
  "us", "ms", "s", "min", "hr"), ..., value.name = NULL)
}
\arguments{
\item{abf}{an abf object.}

\item{intv}{a time interval to melt.}

\item{channel}{channels to melt.}

\item{sample_ratio}{a sampling ratio if need to reduce data points.}

\item{sample_func}{a sampling function if sample_ratio > 1.}

\item{sample_colFunc}{a column vectorised sampling function, used instead of
sample_func for better performance if not NULL.}

\item{time_unit}{a time unit.}

\item{...}{passed to sample_func or sample_colFunc.}

\item{value.name}{a character vector to identity channel columns.}
}
\value{
a data.frame
}
\description{
Melt an abf object into a data.frame grouped by episodes. A time column with
unit provided by time_unit is added to encode time dimension of the orignal
abf object.
}
\details{
For performance consideration, using a column function by sample_colFunc is
strongly encouraged.
}
