% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{MeltAbf}
\alias{MeltAbf}
\title{Melt an abf object.}
\usage{
MeltAbf(abf, intv = NULL, channel = 1L, sampling_ratio = 1L,
  sampling_colFunc = NULL, time_unit = "tick", ...,
  value.name = NULL)
}
\arguments{
\item{abf}{an abf object.}

\item{intv}{a time interval to melt, default is the whole timespan of abj}

\item{channel}{channels to melt.}

\item{sampling_ratio}{sampling ratio.}

\item{sampling_colFunc}{sampling function.}

\item{time_unit}{time unit of the melted data, can be tick, us, ms, s, min or hr.}

\item{...}{further arguments passed to sampling_colFunc}

\item{value.name}{a string to identify/name channel columns, default is named
by channel description without unit.}
}
\value{
a melted data frame.
}
\description{
Melt an abf object into a data.frame grouped by episodes. A time column with
unit provided by time_unit is added to encode time dimension of the orignal
abf object.
}
\details{
For performance consideration, sampling_colFunc should be a vectorised column
function such as colMeans etc, which apply sampling functions to sampled data
points by columns of episodes.
}
