% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{MultiMean}
\alias{MultiMean}
\title{Calculate mean values of multiple abf objects}
\usage{
MultiMean(abf_list, intv_list, channel = 1, na.rm = TRUE)
}
\arguments{
\item{abf_list}{a list of abf objects.}

\item{intv_list}{OPTIONAL, a list of intervals.}

\item{channel}{channel id, 1-based.}

\item{na.rm}{wheter to remove na values..}
}
\value{
A data.frame object.
}
\description{
Calculate mean values of multiple abf objects
}
