% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiPlot.R
\name{MultiPlotChannel}
\alias{MultiPlotChannel}
\title{Plot channel/channels of multiple abf objects.}
\usage{
MultiPlotChannel(abf_list, channel = 1L, intv = NULL, cursor = NULL,
  colour = FALSE, time_unit = "tick", auto_zoom = FALSE,
  manual_title = NULL, auto_title = FALSE, num_label = FALSE,
  channel_arrange = "H", plot_arrange = "A", ...)
}
\arguments{
\item{abf_list}{a list of abf objects.}

\item{channel}{channel/channels to plot, 1-based.}

\item{intv}{OPTIONAL, an interval (a vector of c(start, end, len))/a list of intervals to plot on top.}

\item{cursor}{OPTIONAL, cursors (a vector of positions)/a list of cursors to plot on top.}

\item{colour}{whether to plot in coloured mode.}

\item{time_unit}{time unit for x axis.}

\item{auto_zoom}{whether to zoom in the plot automatically.}

\item{manual_title}{OPTIONAL, titles of every abf plot.}

\item{auto_title}{OPTIONAL, whether to automatically add titles for every abf plot.}

\item{num_label}{OPTIONAL, whether to add number label to every abf plot.}

\item{channel_arrange}{arrangement of channel subplots.}

\item{plot_arrange}{arrangement of abf plots.}

\item{...}{}
}
\value{
a ggplot object.
}
\description{
Plot channel/channels of multiple abf objects.
}
