% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiPlot.R
\name{MultiPlotChannel_Pub}
\alias{MultiPlotChannel_Pub}
\title{Title}
\usage{
MultiPlotChannel_Pub(abf_list, channel = 1, time_unit = "s",
  title_label = FALSE, spacing = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
