% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotChannel.R
\name{PlotChannel}
\alias{PlotChannel}
\title{Plot a channel.}
\usage{
PlotChannel(abf, channel = 1, colour = FALSE, time_unit = "tick",
  ...)
}
\arguments{
\item{abf}{an abf object.}

\item{channel}{channel to plot, channel id is 1-based.}

\item{colour}{wheter to plot in coloured mode.}

\item{time_unit}{time unit for the plot.}

\item{...}{other arguments, see melt.abf for more details.}
}
\value{
a ggplot object.
}
\description{
Plot a channel.
}
