% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotChannel.R
\name{PlotChannel_Cursor}
\alias{PlotChannel_Cursor}
\title{Title}
\usage{
PlotChannel_Cursor(abf, cursor, channel = 1, colour = FALSE,
  time_unit = "tick", ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
