% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotChannel.R
\name{PlotChannel_Cursor}
\alias{PlotChannel_Cursor}
\title{Plot a channel with cursors.}
\usage{
PlotChannel_Cursor(abf, channel = 1L, cursor, colour = FALSE,
  time_unit = "tick", auto_zoom = FALSE, title = NULL, ...)
}
\arguments{
\item{abf}{an abf object.}

\item{channel}{channel to plot, 1-based.}

\item{cursor}{cursors (a vector of positions) to plot on top.}

\item{colour}{whether to plot in coloured mode.}

\item{time_unit}{time unit for x axis.}

\item{auto_zoom}{whether to zoom in the plot automatically.}

\item{title}{OPTIONAL, title of the plot.}

\item{...}{other arguments passed to melt, see melt.abf for more details.}
}
\value{
a ggplot object.
}
\description{
Plot a channel with cursors.
}
