% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotIVSummary}
\alias{PlotIVSummary}
\title{Plot IV Summary generated by IVSummary().}
\usage{
PlotIVSummary(ivs, conductance = "Conductance" \%in\% names(ivs),
  colour = TRUE, title = NULL, legend_title = "",
  error_bar = !conductance, zero_axes = TRUE, smooth = FALSE, ...)
}
\arguments{
\item{ivs}{a data.frame generated by IVSummary or a list of such data.frames.}

\item{conductance}{whether to plot conductance instead of current. Default is
to plot conductance if Conductance is found in}

\item{colour}{whether to plot in colour mode.}

\item{title}{title of the plot.}

\item{legend_title}{title of the legend (group).}

\item{error_bar}{whether to plot error bars. Default is TRUE unless plotting
conductance.}

\item{zero_axes}{wheter to apply ZeroAxes() to plot.}

\item{smooth}{wheter to plot b-Spline smoothed lines.}

\item{...}{not used.}
}
\value{
a ggplot object.
}
\description{
Plot IV Summary generated by IVSummary().
}
