% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuickPlot.R
\name{QuickPlotIV}
\alias{QuickPlotIV}
\title{Quick plot I-V curves at given position}
\usage{
QuickPlotIV(abf, pos, smooth = FALSE)
}
\arguments{
\item{abf}{an abf object.}

\item{pos}{an interval or a cursor.}
}
\value{
a ggplot object.
}
\description{
Quick plot I-V curves at given position
}
