% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuickPlot.R
\name{QuickPlotIV}
\alias{QuickPlotIV}
\title{Quick plot I-V curves at given position}
\usage{
QuickPlotIV(abf, pos)
}
\arguments{
\item{abf}{an abf or a list of abf objects.}

\item{pos}{an interval/cursor or a list of intervals/cursors.}
}
\value{
a ggplot object.
}
\description{
Quick plot I-V curves at given position
}
