% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel.R
\name{ReplaceChannel}
\alias{ReplaceChannel}
\title{Replacing channel data, by-ref like behaviour.}
\usage{
ReplaceChannel(abf, channel, channel_data)
}
\arguments{
\item{channel}{ADC channel id, 1-based.}

\item{channel_data}{channel data to replace the original.}

\item{x}{an abf object.}
}
\value{
an abf object with the replaced channel.
}
\description{
Replacing channel data, by-ref like behaviour.
}
\examples{
#abf itself is changed, no need to assign.
ReplaceChannel(abf, 2, data)
}
