% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode.R
\name{ResEpi}
\alias{ResEpi}
\title{Restore previous removed episodes.}
\usage{
ResEpi(abf, episodes)
}
\arguments{
\item{abf}{an abf object.}

\item{episodes}{episodes to restore.}
}
\value{
an abf object.
}
\description{
Restore previous removed episodes.
}
