% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode.R
\name{RestoreEpisodes}
\alias{RestoreEpisodes}
\title{Restore previous removed episodes, by-ref behaviour.}
\usage{
RestoreEpisodes(abf, episodes)
}
\arguments{
\item{abf}{an abf or a list of abf object.}

\item{episodes}{episodes to restore.}
}
\value{
an abf object.
}
\description{
Restore previous removed episodes, by-ref behaviour.
}
