% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel.R
\name{RplcChan}
\alias{RplcChan}
\title{Replacing channel data.}
\usage{
RplcChan(abf, channel, channel_data)
}
\arguments{
\item{channel}{ADC channel id, 1-based.}

\item{channel_data}{channel data to replace the original.}

\item{x}{an abf object.}
}
\value{
an abf object with the replaced channel.
}
\description{
Replacing channel data.
}
\examples{
abf_new <- RplcChan(abf, 2, data)
}
