% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{SampleAbf}
\alias{SampleAbf}
\title{Sample abf object to reduce data points, by-ref like behaviour.}
\usage{
SampleAbf(abf, sampling_ratio, sampling_func = NULL)
}
\arguments{
\item{abf}{an abf object.}

\item{sampling_ratio}{the sampling ratio. See melt.abf for more details.}

\item{sampling_func}{a sampling function applied to sampled points. See melt.abf for more details.}
}
\value{
the sampled abf itself
}
\description{
Sample abf object to reduce data points, by-ref like behaviour.
}
