% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R
\name{SetChannelUnit}
\alias{SetChannelUnit}
\title{Set channel unit.}
\usage{
SetChannelUnit(abf, unit, channel = 1L)
}
\arguments{
\item{abf}{an abf object.}

\item{unit}{a new unit}

\item{channel}{channel id.}
}
\value{
a abf object with new unit set.
}
\description{
Set channel unit.
}
