% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R
\name{SetTitle}
\alias{SetTitle}
\title{Set title of an abf object, by-ref behaviour.}
\usage{
SetTitle(abf, title)
}
\arguments{
\item{abf}{an abf object, a list of abf objects are also supported.}

\item{title}{the title to be set.}
}
\value{
an abf object, invisibly
}
\description{
Set title of an abf object, by-ref behaviour.
}
