% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{SmplAbf}
\alias{SmplAbf}
\alias{SampleAbf}
\title{Sample abf object to reduce data points.}
\usage{
SmplAbf(abf, sample_ratio, sample_func = NULL, ...)

SampleAbf(abf, sample_ratio, sample_func = NULL, ...)
}
\arguments{
\item{abf}{an abf object.}

\item{sample_ratio}{the sampling ratio. See melt.abf for more details.}

\item{sample_func}{a sampling function applied to sampled points.}

\item{...}{arguments passed to sample_func}
}
\value{
a sampled abf object
}
\description{
Sample abf object to reduce data points.
}
