% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{SmplAbf}
\alias{SmplAbf}
\title{Sample abf object to reduce data points.}
\usage{
SmplAbf(abf, sampling_ratio, sampling_colFunc = NULL, ...)
}
\arguments{
\item{abf}{an abf object.}

\item{sampling_ratio}{the sampling ratio. See melt.abf for more details.}

\item{sampling_colFunc}{a sampling function applied to sampled points.}

\item{...}{arguments passed to sampling_colFunc}
}
\value{
a sampled abf object
}
\description{
Sample abf object to reduce data points.
}
