% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{SmplAbf}
\alias{SmplAbf}
\title{Sampling abf object to reduce data points.}
\usage{
SmplAbf(abf, sampling_ratio, sampling_func = NULL)
}
\arguments{
\item{abf}{an abf object.}

\item{sampling_ratio}{the sampling ratio. See melt.abf for more details.}

\item{sampling_func}{a sampling function applied to sampled points. See melt.abf for more details.}
}
\description{
Sampling abf object to reduce data points.
}
