% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacitance.R
\name{StepMemtestAbf}
\alias{StepMemtestAbf}
\title{Calculate membrane properties from a step pulse membrane test.}
\usage{
StepMemtestAbf(abf, intv = NULL, episode = NULL,
  epoch = FindMemtestEpoch(abf, dac = dac, type = "step"),
  dac = GetWaveformEnabledDAC(abf),
  current_channel = GetFirstCurrentChan(abf), max_iter = 10L)
}
\arguments{
\item{abf}{an abf object.}

\item{intv}{interval of the CHARGING period.}

\item{episode}{episode of the memtest.}

\item{epoch}{epoch of the memtest.}

\item{dac}{DAC channel of Vcmd}

\item{current_channel}{current channel.}

\item{max_iter}{maximum iterations.}
}
\value{
a data.frame
}
\description{
Calculate membrane properties from a step pulse membrane test.
}
\details{
This function calculates membreane properties given a proper step
pulse is applied and recorded in the abf object. Usually all arguments can
be determined automatically. By default episode is determined by
\code{\link[abftools:FindStepEpisode]{FindStepEpisode()}}, see help for more
details.

Please also notice that all unit prefixes are converted to "1", thus the
returned properties calculated are:

Cm: membrane capacitance (in Farad)

Rs: series resistance (in Ohms)

Rm: membrane resistance (in Ohms)

I0: Max charging current (in A)

Is: Steady current (in A)

tao: RC time constant (in s)
}
