% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3func.R
\name{[[<-.abf}
\alias{[[<-.abf}
\title{Replacing channel data.}
\usage{
\method{[[}{abf}(x, channel, channel_data) <- value
}
\arguments{
\item{x}{an abf object.}

\item{channel}{ADC channel id, 1-based.}

\item{channel_data}{channel data to replace the original.}
}
\value{
an abf object with the replaced channel.
}
\description{
Replacing channel data.
}
