% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aux.R
\name{TickToTime}
\alias{TickToTime}
\title{Convert tick (array index) to time unit.}
\usage{
TickToTime(abf, tick, time_unit = c("tick", "us", "ms", "s", "min",
  "hr"), sampling_rate)
}
\arguments{
\item{abf}{an abf object.}

\item{tick}{a vector of integer.}

\item{time_unit}{desired time unit.}

\item{sampling_rate}{sampling rate (in µs).}
}
\value{
a vector of numeric.
}
\description{
Convert tick (array index) to time unit.
}
