% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{wrap_along}
\alias{wrap_along}
\title{Convenient alias of WrapMappingFuncAlong()}
\usage{
wrap_along(..., abf_id_func = NULL, epi_id_func = DefaultEpiLabel,
  chan_id_func = DefaultChanLabel, time_unit = "ms", ret.df = TRUE)
}
\arguments{
\item{...}{See WrapMappingFuncAlong for more details.}
}
\value{
a function of f(abf, intv, episode, channel)
}
\description{
Convenient alias of WrapMappingFuncAlong()
}
