% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_components.R
\name{ZeroAxes}
\alias{ZeroAxes}
\alias{ApplyZeroAxes}
\title{Draw axes centred at zero}
\usage{
ZeroAxes(xrange, yrange, xlabel, ylabel, xticks = 5, yticks = 5,
  basesize = 5)

ApplyZeroAxes(p, xticks = 5, yticks = 5, basesize = 5)
}
\arguments{
\item{xrange}{Range of x axis.}

\item{yrange}{Range of y axis.}

\item{xlabel}{Label of x axis.}

\item{ylabel}{Label of y axis.}

\item{xticks}{Number of ticks on x axis.}

\item{yticks}{Number of ticks on y axis.}

\item{basesize}{Base textsize.}

\item{p}{A ggplot object.}
}
\value{
a list of ggplot objects.
}
\description{
Draw axes centred at zero
}
\details{
ZerosAxes() draw axes by annotating on (0, ) and (, 0). Since annotate()
is used, theme settings are not inherited, thus basesize must be set independently.
}
